Re-number the PBR Experiments
Now that we're adding metal platform we need to re-number all the experiments so we can add the new platform.
