+++ assets/projects/202_pbr_camera/drawtexture.hlsl
