+++ assets/projects/251_pbr_explorer/drawtexture.hlsl
