+++ assets/projects/252_pbr_material_properties/drawtexture.hlsl
