+++ assets/projects/253_pbr_material_textures/drawtexture.hlsl
