+++ projects/pbr/201_pbr_spheres_d3d12/201_pbr_spheres_d3d12.cpp
@@ -126,7 +126,7 @@ int main(int argc, char** argv)
     std::vector<char> dxilVS;
     std::vector<char> dxilPS;
     {
-        std::string shaderSource = LoadString("projects/201_202_pbr_spheres/shaders.hlsl");
+        std::string shaderSource = LoadString("projects/201_pbr_spheres/shaders.hlsl");
 
         std::string errorMsg;
         HRESULT     hr = CompileHLSL(shaderSource, "vsmain", "vs_6_0", &dxilVS, &errorMsg);
@@ -153,7 +153,7 @@ int main(int argc, char** argv)
     std::vector<char> drawTextureDxilVS;
     std::vector<char> drawTextureDxilPS;
     {
-        std::string shaderSource = LoadString("projects/201_202_pbr_spheres/drawtexture.hlsl");
+        std::string shaderSource = LoadString("projects/201_pbr_spheres/drawtexture.hlsl");
         if (shaderSource.empty()) {
             assert(false && "no shader source");
             return EXIT_FAILURE;
