+++ projects/pbr/201_pbr_spheres_vulkan/201_pbr_spheres_vulkan.cpp
@@ -152,7 +152,7 @@ int main(int argc, char** argv)
     std::vector<uint32_t> spirvVS;
     std::vector<uint32_t> spirvFS;
     {
-        std::string shaderSource = LoadString("projects/201_202_pbr_spheres/shaders.hlsl");
+        std::string shaderSource = LoadString("projects/201_pbr_spheres/shaders.hlsl");
 
         std::string errorMsg;
         HRESULT     hr = CompileHLSL(shaderSource, "vsmain", "vs_6_0", &spirvVS, &errorMsg);
@@ -198,7 +198,7 @@ int main(int argc, char** argv)
     std::vector<uint32_t> drawTextureSpirvVS;
     std::vector<uint32_t> drawTextureSpirvFS;
     {
-        std::string shaderSource = LoadString("projects/201_202_pbr_spheres/drawtexture.hlsl");
+        std::string shaderSource = LoadString("projects/201_pbr_spheres/drawtexture.hlsl");
         if (shaderSource.empty()) {
             assert(false && "no shader source");
             return EXIT_FAILURE;
