+++ projects/pbr/203_pbr_align_vulkan/203_pbr_align_vulkan.cpp
@@ -152,7 +152,7 @@ int main(int argc, char** argv)
     std::vector<uint32_t> spirvVS;
     std::vector<uint32_t> spirvFS;
     {
-        std::string shaderSource = LoadString("projects/205_206_pbr_align/shaders.hlsl");
+        std::string shaderSource = LoadString("projects/203_pbr_align/shaders.hlsl");
 
         std::string errorMsg;
         HRESULT     hr = CompileHLSL(shaderSource, "vsmain", "vs_6_0", &spirvVS, &errorMsg);
@@ -198,7 +198,7 @@ int main(int argc, char** argv)
     std::vector<uint32_t> drawTextureSpirvVS;
     std::vector<uint32_t> drawTextureSpirvFS;
     {
-        std::string shaderSource = LoadString("projects/205_206_pbr_align/drawtexture.hlsl");
+        std::string shaderSource = LoadString("projects/203_pbr_align/drawtexture.hlsl");
         if (shaderSource.empty()) {
             assert(false && "no shader source");
             return EXIT_FAILURE;
