+++ projects/pbr/251_pbr_explorer_vulkan/251_pbr_explorer_vulkan.cpp
@@ -10,8 +10,6 @@
 #include <glm/gtx/transform.hpp>
 using namespace glm;
 
-#include <fstream>
-
 #define CHECK_CALL(FN)                               \
     {                                                \
         HRESULT hr = FN;                             \
@@ -26,21 +24,32 @@ using namespace glm;
         }                                            \
     }
 
-#define MATERIAL_TEXTURE_STRIDE 4
-#define NUM_MATERIALS           16 
-#define TOTAL_MATERIAL_TEXTURES (NUM_MATERIALS * MATERIAL_TEXTURE_STRIDE)
+#define DISTRIBUTION_TROWBRIDGE_REITZ 0
+#define DISTRIBUTION_BECKMANN         1
+#define DISTRIBUTION_BLINN_PHONG      2
 
-#define IBL_INTEGRATION_LUT_DESCRIPTOR_OFFSET    3
-#define IBL_INTEGRATION_MS_LUT_DESCRIPTOR_OFFSET 4
-#define IBL_IRRADIANCE_MAPS_DESCRIPTOR_OFFSET    16
-#define IBL_ENVIRONMENT_MAPS_DESCRIPTOR_OFFSET   48
-#define MATERIAL_TEXTURES_DESCRIPTOR_OFFSET      100
+#define FRESNEL_SCHLICK_ROUGHNESS 0
+#define FRESNEL_SCHLICK           1
+#define FRESNEL_COOK_TORRANCE     2
+#define FRESNEL_NONE              3
+
+#define GEOMETRY_SMITH                 0
+#define GEOMETRY_IMPLICIT              1
+#define GEOMETRY_NEUMANN               2
+#define GEOMETRY_COOK_TORRANCE         3
+#define GEOMETRY_KELEMEN               4
+#define GEOMETRY_BECKMANN              5
+#define GEOMETRY_GGX1                  6
+#define GEOMETRY_GGX2                  7
+#define GEOMETRY_SCHLICK_GGX           8
+#define GEOMETRY_SMITH_CORRELATED      9
+#define GEOMETRY_SMITH_CORRELATED_FAST 10
 
 // This will be passed in via constant buffer
 struct Light
 {
-    uint32_t active;
     vec3     position;
+    uint32_t __pad;
     vec3     color;
     float    intensity;
 };
@@ -53,34 +62,35 @@ struct PBRSceneParameters
     Light    lights[8];
     uint32_t iblNumEnvLevels;
     uint32_t iblIndex;
-    uint     multiscatter;
-    uint     colorCorrect;
+    float    iblDiffuseStrength;
+    float    iblSpecularStrength;
 };
 
 struct EnvSceneParameters
 {
-    mat4     MVP;
-    uint32_t IBLIndex;
-};
-
-struct MaterialParameters
-{
-    float specular;
+    mat4 MVP;
+    uint IBLIndex;
 };
 
 struct DrawParameters
 {
-    mat4     ModelMatrix;
-    uint32_t MaterialIndex;
-    uint32_t InvertNormalMapY;
+    mat4 ModelMatrix;
+    uint MaterialIndex;
 };
 
-struct MaterialTextures
+struct MaterialParameters
 {
-    VulkanImage baseColorTexture;
-    VulkanImage normalTexture;
-    VulkanImage roughnessTexture;
-    VulkanImage metallicTexture;
+    vec3     baseColor;
+    float    roughness;
+    float    metallic;
+    float    specular;
+    uint     directComponentMode;
+    uint32_t D_Func;
+    uint32_t F_Func;
+    uint32_t G_Func;
+    uint32_t indirectComponentMode;
+    uint32_t indirectSpecularMode;
+    uint32_t drawMode;
 };
 
 struct GeometryBuffers
@@ -98,11 +108,68 @@ struct GeometryBuffers
 // Constants
 // =============================================================================
 
+const std::vector<std::string> gDistributionNames = {
+    "GGX (Trowbridge-Reitz)",
+    "Beckmann",
+    "Blinn-Phong",
+};
+
+const std::vector<std::string> gFresnelNames = {
+    "Schlick with Roughness",
+    "Schlick",
+    "CookTorrance",
+    "None",
+};
+
+const std::vector<std::string> gGeometryNames = {
+    "Smith",
+    "Implicit",
+    "Neumann",
+    "Cook-Torrance",
+    "Kelemen",
+    "Beckmann",
+    "GGX1",
+    "GGX2",
+    "SchlickGGX",
+    "Smith Correlated",
+    "Smith Correlated Fast",
+};
+
+const std::vector<std::string> gDirectComponentModeNames = {
+    "All",
+    "Distribution",
+    "Fresnel",
+    "Geometry",
+    "Diffuse",
+    "Radiance",
+    "kD",
+    "Specular",
+    "BRDF",
+};
+
+const std::vector<std::string> gIndirectComponentModeNames = {
+    "All",
+    "Diffuse",
+    "Specular"};
+
+const std::vector<std::string> gIndirectSpecularModeNames = {
+    "LUT",
+    "Approx Lazarov",
+    "Approx Polynomial",
+    "Approx Karis",
+};
+
+const std::vector<std::string> gDrawModeNames = {
+    "Full Lighting",
+    "Direct",
+    "Indirect",
+};
+
 const std::vector<std::string> gModelNames = {
     "Sphere",
     "Knob",
     "Monkey",
-    "Cube",
+    "Teapot",
 };
 
 // =============================================================================
@@ -113,16 +180,45 @@ static uint32_t gWindowHeight     = 1080;
 static bool     gEnableDebug      = true;
 static bool     gEnableRayTracing = false;
 
+static LPCWSTR gVSShaderName = L"vsmain";
+static LPCWSTR gPSShaderName = L"psmain";
+
 static float gTargetAngle = 0.0f;
 static float gAngle       = 0.0f;
 
-static std::vector<std::string> gMaterialNames = {};
+// clang-format off
+static std::vector<MaterialParameters> gMaterialParams = {
+    {F0_MetalCopper,         0.25f, 1.00f, 0.5f, 0, 0, 0, 0},
+    {F0_MetalGold,           0.05f, 1.00f, 0.5f, 0, 0, 0, 0},
+    {F0_MetalSilver,         0.18f, 1.00f, 0.5f, 0, 0, 0, 0},
+    {F0_MetalZinc,           0.65f, 1.00f, 0.5f, 0, 0, 0, 0},
+    {F0_MetalTitanium,       0.11f, 1.00f, 0.5f, 0, 0, 0, 0},
+    {vec3(0.6f, 0.0f, 0.0f), 0.00f, 0.00f, 0.5f, 0, 0, 0, 0},
+    {vec3(0.0f, 0.6f, 0.0f), 0.25f, 0.00f, 0.5f, 0, 0, 0, 0},
+    {vec3(0.0f, 0.0f, 0.6f), 0.50f, 0.00f, 0.5f, 0, 0, 0, 0},
+    {vec3(0.7f, 0.7f, 0.2f), 0.92f, 0.15f, 0.5f, 0, 0, 0, 0},
+};
+// clang-format on
 
-static uint32_t                 gNumLights  = 4;
-static const uint32_t           gMaxIBLs    = 32;
-static uint32_t                 gIBLIndex   = 0;
-static std::vector<std::string> gIBLNames   = {};
-static uint32_t                 gModelIndex = 0;
+static std::vector<std::string> gMaterialNames = {
+    "Copper",
+    "Gold",
+    "Silver",
+    "Zink",
+    "Titanium",
+    "Shiny Plastic",
+    "Rough Plastic",
+    "Rougher Plastic",
+    "Roughest Plastic",
+};
+
+static uint32_t                 gNumLights           = 0;
+static const uint32_t           gMaxIBLs             = 32;
+static uint32_t                 gIBLIndex            = 0;
+static std::vector<std::string> gIBLNames            = {};
+static float                    gIBLDiffuseStrength  = 1.0f;
+static float                    gIBLSpecularStrength = 1.0f;
+static uint32_t                 gModelIndex          = 0;
 
 void CreatePBRPipeline(VulkanRenderer* pRenderer, VulkanPipelineLayout* pLayout);
 void CreateEnvironmentPipeline(VulkanRenderer* pRenderer, VulkanPipelineLayout* pLayout);
@@ -134,36 +230,36 @@ void CreateMaterialModels(
     std::vector<GeometryBuffers>& outGeomtryBuffers);
 void CreateIBLTextures(
     VulkanRenderer*           pRenderer,
-    VulkanImage*              ppBRDFLUT,
-    VulkanImage*              ppMultiscatterBRDFLUT,
+    VulkanImage*              pBRDFLUT,
     std::vector<VulkanImage>& outIrradianceTextures,
     std::vector<VulkanImage>& outEnvironmentTextures,
     std::vector<uint32_t>&    outEnvNumLevels);
-void CreateMaterials(
-    VulkanRenderer*                  pRenderer,
-    MaterialTextures&                outDefaultMaterialTextures,
-    std::vector<MaterialTextures>&   outMaterialTexturesSets,
-    std::vector<MaterialParameters>& outMaterialParametersSets);
 void CreateDescriptorBuffer(
     VulkanRenderer*       pRenderer,
-    VkDescriptorSetLayout pDescriptorSetLayout,
+    VkDescriptorSetLayout descriptorSetLayout,
     VulkanBuffer*         pBuffer);
 void WritePBRDescriptors(
-    VulkanRenderer*                pRenderer,
-    VkDescriptorSetLayout          descriptorSetLayout,
-    VulkanBuffer*                  pDescriptorBuffer,
-    VulkanBuffer*                  pSceneParamsBuffer,
-    VulkanBuffer*                  pMaterialBuffer,
-    std::vector<MaterialTextures>& materialTextureSets,
-    const VulkanImage*             pBRDFLUT,
-    const VulkanImage*             pMultiscatterBRDFLUT,
-    std::vector<VulkanImage>&      irrTextures,
-    std::vector<VulkanImage>&      envTextures);
+    VulkanRenderer*           pRenderer,
+    VkDescriptorSetLayout     descriptorSetLayout,
+    VulkanBuffer*             pDescriptorBuffer,
+    const VulkanBuffer*       pSceneParamsBuffer,
+    const VulkanBuffer*       pMaterialsBuffer,
+    const VulkanImage*        pBRDFLUT,
+    std::vector<VulkanImage>& pIrradianceTexture,
+    std::vector<VulkanImage>& pEnvTexture);
 void WriteEnvDescriptors(
-    VulkanRenderer*          pRenderer,
-    VkDescriptorSetLayout    descriptorSetLayout,
-    VulkanBuffer*            pDescriptorBuffer,
-    std::vector<VulkanImage> envTextures);
+    VulkanRenderer*           pRenderer,
+    VkDescriptorSetLayout     descriptorSetLayout,
+    VulkanBuffer*             pDescriptorBuffer,
+    std::vector<VulkanImage>& envTextures);
+
+/*
+void CreatePBRRootSig(DxRenderer* pRenderer, ID3D12RootSignature** ppRootSig);
+void CreateEnvironmentRootSig(DxRenderer* pRenderer, ID3D12RootSignature** ppRootSig);
+void CreateDescriptorHeap(
+    DxRenderer*            pRenderer,
+    ID3D12DescriptorHeap** ppHeap);
+    */
 
 void MouseMove(int x, int y, int buttons)
 {
@@ -199,7 +295,7 @@ int main(int argc, char** argv)
     std::vector<uint32_t> spirvVS;
     std::vector<uint32_t> spirvFS;
     {
-        std::string shaderSource = LoadString("projects/255_256_pbr_material_textures/shaders.hlsl");
+        std::string shaderSource = LoadString("projects/251_pbr_explorer/shaders.hlsl");
         if (shaderSource.empty()) {
             assert(false && "no shader source");
             return EXIT_FAILURE;
@@ -249,7 +345,7 @@ int main(int argc, char** argv)
     std::vector<uint32_t> drawTextureSpirvVS;
     std::vector<uint32_t> drawTextureSpirvFS;
     {
-        std::string shaderSource = LoadString("projects/255_256_pbr_material_textures/drawtexture.hlsl");
+        std::string shaderSource = LoadString("projects/251_pbr_explorer/drawtexture.hlsl");
         if (shaderSource.empty()) {
             assert(false && "no shader source");
             return EXIT_FAILURE;
@@ -310,8 +406,8 @@ int main(int argc, char** argv)
     // *************************************************************************
     // PBR pipeline state object
     // *************************************************************************
-    VkPipeline pbrPipelineState;
-    CHECK_CALL(CreateGraphicsPipeline1(
+    VkPipeline pbrPipelineState = VK_NULL_HANDLE;
+    CHECK_CALL(CreateDrawNormalPipeline(
         renderer.get(),
         pbrPipelineLayout.PipelineLayout,
         shaderModuleVS,
@@ -319,12 +415,15 @@ int main(int argc, char** argv)
         GREX_DEFAULT_RTV_FORMAT,
         GREX_DEFAULT_DSV_FORMAT,
         &pbrPipelineState,
-        VK_CULL_MODE_BACK_BIT));
+        false, // enable tangents
+        VK_CULL_MODE_BACK_BIT,
+        "vsmain",
+        "psmain"));
 
     // *************************************************************************
     // Environment pipeline state object
     // *************************************************************************
-    VkPipeline envPipelineState;
+    VkPipeline envPipelineState = VK_NULL_HANDLE;
     CHECK_CALL(CreateDrawTexturePipeline(
         renderer.get(),
         envPipelineLayout.PipelineLayout,
@@ -338,7 +437,20 @@ int main(int argc, char** argv)
         "psmain"));
 
     // *************************************************************************
-    // Constant buffer
+    // Material buffer
+    // *************************************************************************
+    VulkanBuffer pbrMaterialParamsBuffer = {};
+    CHECK_CALL(CreateBuffer(
+        renderer.get(),
+        SizeInBytes(gMaterialParams),
+        DataPtr(gMaterialParams),
+        VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT | VK_BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT,
+        VMA_MEMORY_USAGE_CPU_TO_GPU,
+        0,
+        &pbrMaterialParamsBuffer));
+
+    // *************************************************************************
+    // Constant buffers
     // *************************************************************************
     VulkanBuffer pbrSceneParamsBuffer;
     CHECK_CALL(CreateBuffer(
@@ -370,42 +482,10 @@ int main(int argc, char** argv)
     // Environment texture
     // *************************************************************************
     VulkanImage              brdfLUT;
-    VulkanImage              multiscatterBRDFLUT;
     std::vector<VulkanImage> irrTextures;
     std::vector<VulkanImage> envTextures;
     std::vector<uint32_t>    envNumLevels;
-    CreateIBLTextures(
-        renderer.get(),
-        &brdfLUT,
-        &multiscatterBRDFLUT,
-        irrTextures,
-        envTextures,
-        envNumLevels);
-
-    // *************************************************************************
-    // Material texture
-    // *************************************************************************
-    MaterialTextures                defaultMaterialTextures;
-    std::vector<MaterialTextures>   materialTexturesSets;
-    std::vector<MaterialParameters> materialParametersSets;
-    CreateMaterials(
-        renderer.get(),
-        defaultMaterialTextures,
-        materialTexturesSets,
-        materialParametersSets);
-
-    // *************************************************************************
-    // Material buffer
-    // *************************************************************************
-    VulkanBuffer materialBuffer;
-    CHECK_CALL(CreateBuffer(
-        renderer.get(),
-        SizeInBytes(materialParametersSets),
-        DataPtr(materialParametersSets),
-        VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT | VK_BUFFER_USAGE_SHADER_DEVICE_ADDRESS_BIT,
-        VMA_MEMORY_USAGE_CPU_TO_GPU,
-        0,
-        &materialBuffer));
+    CreateIBLTextures(renderer.get(), &brdfLUT, irrTextures, envTextures, envNumLevels);
 
     // *************************************************************************
     // Descriptor buffers
@@ -418,10 +498,8 @@ int main(int argc, char** argv)
         pbrPipelineLayout.DescriptorSetLayout,
         &pbrDescriptorBuffer,
         &pbrSceneParamsBuffer,
-        &materialBuffer,
-        materialTexturesSets,
+        &pbrMaterialParamsBuffer,
         &brdfLUT,
-        &multiscatterBRDFLUT,
         irrTextures,
         envTextures);
 
@@ -437,7 +515,7 @@ int main(int argc, char** argv)
     // *************************************************************************
     // Window
     // *************************************************************************
-    auto window = Window::Create(gWindowWidth, gWindowHeight, "256_pbr_material_textures_vulkan");
+    auto window = Window::Create(gWindowWidth, gWindowHeight, "252_pbr_explorer_vulkan");
     if (!window) {
         assert(false && "Window::Create failed");
         return EXIT_FAILURE;
@@ -534,49 +612,48 @@ int main(int argc, char** argv)
     }
 
     // *************************************************************************
-    // Persistent map parameters
+    // Persistent map scene parameters
     // *************************************************************************
     PBRSceneParameters* pPBRSceneParams = nullptr;
     vmaMapMemory(renderer->Allocator, pbrSceneParamsBuffer.Allocation, reinterpret_cast<void**>(&pPBRSceneParams));
 
     MaterialParameters* pMaterialParams = nullptr;
-    vmaMapMemory(renderer->Allocator, materialBuffer.Allocation, reinterpret_cast<void**>(&pMaterialParams));
-
-    // *************************************************************************
-    // Set some scene params
-    // *************************************************************************
-    pPBRSceneParams->numLights           = gNumLights;
-    pPBRSceneParams->lights[0].active    = 0;
-    pPBRSceneParams->lights[0].position  = vec3(3, 10, 0);
-    pPBRSceneParams->lights[0].color     = vec3(1, 1, 1);
-    pPBRSceneParams->lights[0].intensity = 1.5f;
-    pPBRSceneParams->lights[1].active    = 0;
-    pPBRSceneParams->lights[1].position  = vec3(-8, 1, 4);
-    pPBRSceneParams->lights[1].color     = vec3(0.85f, 0.95f, 0.81f);
-    pPBRSceneParams->lights[1].intensity = 0.4f;
-    pPBRSceneParams->lights[2].active    = 0;
-    pPBRSceneParams->lights[2].position  = vec3(0, 8, -8);
-    pPBRSceneParams->lights[2].color     = vec3(0.89f, 0.89f, 0.97f);
-    pPBRSceneParams->lights[2].intensity = 0.95f;
-    pPBRSceneParams->lights[3].active    = 0;
-    pPBRSceneParams->lights[3].position  = vec3(15, 0, 0);
-    pPBRSceneParams->lights[3].color     = vec3(0.92f, 0.5f, 0.7f);
-    pPBRSceneParams->lights[3].intensity = 0.5f;
-    pPBRSceneParams->iblNumEnvLevels     = envNumLevels[gIBLIndex];
-    pPBRSceneParams->iblIndex            = gIBLIndex;
-    pPBRSceneParams->colorCorrect        = 0;
+    vmaMapMemory(renderer->Allocator, pbrMaterialParamsBuffer.Allocation, reinterpret_cast<void**>(&pMaterialParams));
 
     // *************************************************************************
     // Main loop
     // *************************************************************************
     VkClearValue clearValues[2];
-    clearValues[0].color        = { {0.0f, 0.0f, 0.2f, 1.0f} };
+    clearValues[0].color = {
+        {0.0f, 0.0f, 0.2f, 1.0f}
+    };
     clearValues[1].depthStencil = {1.0f, 0};
 
     while (window->PollEvents()) {
         window->ImGuiNewFrameVulkan();
 
         if (ImGui::Begin("Scene")) {
+            static const char* currentIBLName = gIBLNames[0].c_str();
+            if (ImGui::BeginCombo("IBL", currentIBLName)) {
+                for (size_t i = 0; i < gIBLNames.size(); ++i) {
+                    bool isSelected = (currentIBLName == gIBLNames[i]);
+                    if (ImGui::Selectable(gIBLNames[i].c_str(), isSelected)) {
+                        currentIBLName = gIBLNames[i].c_str();
+                        gIBLIndex      = static_cast<uint32_t>(i);
+                    }
+                    if (isSelected) {
+                        ImGui::SetItemDefaultFocus();
+                    }
+                }
+                ImGui::EndCombo();
+            }
+
+            ImGui::SliderFloat("IBL Diffuse Strength", &gIBLDiffuseStrength, 0.0f, 2.0f);
+            ImGui::SliderFloat("IBL Specular Strength", &gIBLSpecularStrength, 0.0f, 2.0f);
+            ImGui::SliderInt("Number of Lights", reinterpret_cast<int*>(&gNumLights), 0, 4);
+
+            ImGui::Separator();
+
             static const char* currentModelName = gModelNames[0].c_str();
             if (ImGui::BeginCombo("Model", currentModelName)) {
                 for (size_t i = 0; i < gModelNames.size(); ++i) {
@@ -591,52 +668,133 @@ int main(int argc, char** argv)
                 }
                 ImGui::EndCombo();
             }
-
-            ImGui::Separator();
-
-            static const char* currentIBLName = gIBLNames[0].c_str();
-            if (ImGui::BeginCombo("IBL", currentIBLName)) {
-                for (size_t i = 0; i < gIBLNames.size(); ++i) {
-                    bool isSelected = (currentIBLName == gIBLNames[i]);
-                    if (ImGui::Selectable(gIBLNames[i].c_str(), isSelected)) {
-                        currentIBLName         = gIBLNames[i].c_str();
-                        pPBRSceneParams->iblIndex = static_cast<uint32_t>(i);
-                    }
-                    if (isSelected) {
-                        ImGui::SetItemDefaultFocus();
-                    }
-                }
-                ImGui::EndCombo();
-            }
-
-            ImGui::Separator();
-
-            ImGui::Checkbox("Multiscatter", reinterpret_cast<bool*>(&pPBRSceneParams->multiscatter));
-
-            ImGui::Separator();
-
-            ImGui::Checkbox("Color Correct", reinterpret_cast<bool*>(&pPBRSceneParams->colorCorrect));
-
-            ImGui::Separator();
-
-            for (uint32_t lightIdx = 0; lightIdx < 4; ++lightIdx) {
-                std::stringstream lightName;
-                lightName << "Light " << lightIdx;
-                if (ImGui::TreeNodeEx(lightName.str().c_str(), ImGuiTreeNodeFlags_None)) {
-                    ImGui::Checkbox("Active", reinterpret_cast<bool*>(&pPBRSceneParams->lights[lightIdx].active));
-                    ImGui::SliderFloat("Intensity", &pPBRSceneParams->lights[lightIdx].intensity, 0.0f, 10.0f);
-                    ImGui::ColorPicker3("Albedo", reinterpret_cast<float*>(&(pPBRSceneParams->lights[lightIdx].color)), ImGuiColorEditFlags_NoInputs);
-
-                    ImGui::TreePop();
-                }
-            }
         }
         ImGui::End();
 
         if (ImGui::Begin("Material Parameters")) {
+            static std::vector<const char*> currentDrawModeNames(gMaterialParams.size(), gDrawModeNames[0].c_str());
+            static std::vector<const char*> currentDirectComponentModeNames(gMaterialParams.size(), gDirectComponentModeNames[0].c_str());
+            static std::vector<const char*> currentDistributionNames(gMaterialParams.size(), gDistributionNames[0].c_str());
+            static std::vector<const char*> currentFresnelNames(gMaterialParams.size(), gFresnelNames[0].c_str());
+            static std::vector<const char*> currentGeometryNames(gMaterialParams.size(), gGeometryNames[0].c_str());
+            static std::vector<const char*> currentIndirectComponentModeNames(gMaterialParams.size(), gIndirectComponentModeNames[0].c_str());
+            static std::vector<const char*> currentIndirectSpecularModeNames(gMaterialParams.size(), gIndirectSpecularModeNames[0].c_str());
+
             for (uint32_t matIdx = 0; matIdx < gMaterialNames.size(); ++matIdx) {
                 if (ImGui::TreeNodeEx(gMaterialNames[matIdx].c_str(), ImGuiTreeNodeFlags_DefaultOpen)) {
+                    // DrawMode
+                    if (ImGui::BeginCombo("DrawMode", currentDrawModeNames[matIdx])) {
+                        for (size_t i = 0; i < gDrawModeNames.size(); ++i) {
+                            bool isSelected = (currentDrawModeNames[matIdx] == gDrawModeNames[i]);
+                            if (ImGui::Selectable(gDrawModeNames[i].c_str(), isSelected)) {
+                                currentDrawModeNames[matIdx]     = gDrawModeNames[i].c_str();
+                                pMaterialParams[matIdx].drawMode = static_cast<uint32_t>(i);
+                            }
+                            if (isSelected) {
+                                ImGui::SetItemDefaultFocus();
+                            }
+                        }
+                        ImGui::EndCombo();
+                    }
+                    // Direct Light Params
+                    if (ImGui::TreeNodeEx("Direct Light Parames", ImGuiTreeNodeFlags_DefaultOpen)) {
+                        // DirectComponentMode
+                        if (ImGui::BeginCombo("Direct Component Mode", currentDirectComponentModeNames[matIdx])) {
+                            for (size_t i = 0; i < gDirectComponentModeNames.size(); ++i) {
+                                bool isSelected = (currentDirectComponentModeNames[matIdx] == gDirectComponentModeNames[i]);
+                                if (ImGui::Selectable(gDirectComponentModeNames[i].c_str(), isSelected)) {
+                                    currentDirectComponentModeNames[matIdx]     = gDirectComponentModeNames[i].c_str();
+                                    pMaterialParams[matIdx].directComponentMode = static_cast<uint32_t>(i);
+                                }
+                                if (isSelected) {
+                                    ImGui::SetItemDefaultFocus();
+                                }
+                            }
+                            ImGui::EndCombo();
+                        }
+                        // Distribution
+                        if (ImGui::BeginCombo("Distribution", currentDistributionNames[matIdx])) {
+                            for (size_t i = 0; i < gDistributionNames.size(); ++i) {
+                                bool isSelected = (currentDistributionNames[matIdx] == gDistributionNames[i]);
+                                if (ImGui::Selectable(gDistributionNames[i].c_str(), isSelected)) {
+                                    currentDistributionNames[matIdx] = gDistributionNames[i].c_str();
+                                    pMaterialParams[matIdx].D_Func   = static_cast<uint32_t>(i);
+                                }
+                                if (isSelected) {
+                                    ImGui::SetItemDefaultFocus();
+                                }
+                            }
+                            ImGui::EndCombo();
+                        }
+                        // Fresnel
+                        if (ImGui::BeginCombo("Fresnel", currentFresnelNames[matIdx])) {
+                            for (size_t i = 0; i < gFresnelNames.size(); ++i) {
+                                bool isSelected = (currentFresnelNames[matIdx] == gFresnelNames[i]);
+                                if (ImGui::Selectable(gFresnelNames[i].c_str(), isSelected)) {
+                                    currentFresnelNames[matIdx]    = gFresnelNames[i].c_str();
+                                    pMaterialParams[matIdx].F_Func = static_cast<uint32_t>(i);
+                                }
+                                if (isSelected) {
+                                    ImGui::SetItemDefaultFocus();
+                                }
+                            }
+                            ImGui::EndCombo();
+                        }
+                        // Geometry
+                        if (ImGui::BeginCombo("Geometry", currentGeometryNames[matIdx])) {
+                            for (size_t i = 0; i < gGeometryNames.size(); ++i) {
+                                bool isSelected = (currentGeometryNames[matIdx] == gGeometryNames[i]);
+                                if (ImGui::Selectable(gGeometryNames[i].c_str(), isSelected)) {
+                                    currentGeometryNames[matIdx]   = gGeometryNames[i].c_str();
+                                    pMaterialParams[matIdx].G_Func = static_cast<uint32_t>(i);
+                                }
+                                if (isSelected) {
+                                    ImGui::SetItemDefaultFocus();
+                                }
+                            }
+                            ImGui::EndCombo();
+                        }
+
+                        ImGui::TreePop();
+                    }
+                    // Indirect Light Params
+                    if (ImGui::TreeNodeEx("Indirect Light Parames", ImGuiTreeNodeFlags_DefaultOpen)) {
+                        // IndirectComponentMode
+                        if (ImGui::BeginCombo("Indirect Component Mode", currentIndirectComponentModeNames[matIdx])) {
+                            for (size_t i = 0; i < gIndirectComponentModeNames.size(); ++i) {
+                                bool isSelected = (currentIndirectComponentModeNames[matIdx] == gIndirectComponentModeNames[i]);
+                                if (ImGui::Selectable(gIndirectComponentModeNames[i].c_str(), isSelected)) {
+                                    currentIndirectComponentModeNames[matIdx]     = gIndirectComponentModeNames[i].c_str();
+                                    pMaterialParams[matIdx].indirectComponentMode = static_cast<uint32_t>(i);
+                                }
+                                if (isSelected) {
+                                    ImGui::SetItemDefaultFocus();
+                                }
+                            }
+                            ImGui::EndCombo();
+                        }
+                        // Specular Mode
+                        if (ImGui::BeginCombo("Specular Mode", currentIndirectSpecularModeNames[matIdx])) {
+                            for (size_t i = 0; i < gIndirectSpecularModeNames.size(); ++i) {
+                                bool isSelected = (currentIndirectSpecularModeNames[matIdx] == gIndirectSpecularModeNames[i]);
+                                if (ImGui::Selectable(gIndirectSpecularModeNames[i].c_str(), isSelected)) {
+                                    currentIndirectSpecularModeNames[matIdx]     = gIndirectSpecularModeNames[i].c_str();
+                                    pMaterialParams[matIdx].indirectSpecularMode = static_cast<uint32_t>(i);
+                                }
+                                if (isSelected) {
+                                    ImGui::SetItemDefaultFocus();
+                                }
+                            }
+                            ImGui::EndCombo();
+                        }
+
+                        ImGui::TreePop();
+                    }
+
+                    ImGui::SliderFloat("Roughness", &(pMaterialParams[matIdx].roughness), 0.0f, 1.0f);
+                    ImGui::SliderFloat("Metallic", &(pMaterialParams[matIdx].metallic), 0.0f, 1.0f);
                     ImGui::SliderFloat("Specular", &(pMaterialParams[matIdx].specular), 0.0f, 1.0f);
+                    ImGui::ColorPicker3("Albedo", reinterpret_cast<float*>(&(pMaterialParams[matIdx].baseColor)), ImGuiColorEditFlags_NoInputs);
 
                     ImGui::TreePop();
                 }
@@ -703,15 +861,34 @@ int main(int argc, char** argv)
 
             // Camera matrices - spin the camera around the target
             mat4 transformEyeMat     = glm::rotate(glm::radians(-gAngle), vec3(0, 1, 0));
-            vec3 startingEyePosition = vec3(0, 2.5f, 10);
+            vec3 startingEyePosition = vec3(0, 3, 8);
             vec3 eyePosition         = transformEyeMat * vec4(startingEyePosition, 1);
             mat4 viewMat             = glm::lookAt(eyePosition, vec3(0, 0, 0), vec3(0, 1, 0));
             mat4 projMat             = glm::perspective(glm::radians(60.0f), gWindowWidth / static_cast<float>(gWindowHeight), 0.1f, 10000.0f);
 
-            // Set scene params values that required calculation
+            // Set constant buffer values
+            //
+            // We're rotating everything in the world...including the lights
+            //
             pPBRSceneParams->viewProjectionMatrix = projMat * viewMat;
             pPBRSceneParams->eyePosition          = eyePosition;
+            pPBRSceneParams->numLights            = gNumLights;
+            pPBRSceneParams->lights[0].position   = vec3(3, 10, 0);
+            pPBRSceneParams->lights[0].color      = vec3(1, 1, 1);
+            pPBRSceneParams->lights[0].intensity  = 1.5f;
+            pPBRSceneParams->lights[1].position   = vec3(-8, 1, 4);
+            pPBRSceneParams->lights[1].color      = vec3(0.85f, 0.95f, 0.81f);
+            pPBRSceneParams->lights[1].intensity  = 0.4f;
+            pPBRSceneParams->lights[2].position   = vec3(0, 8, -8);
+            pPBRSceneParams->lights[2].color      = vec3(0.89f, 0.89f, 0.97f);
+            pPBRSceneParams->lights[2].intensity  = 0.95f;
+            pPBRSceneParams->lights[3].position   = vec3(15, 0, 0);
+            pPBRSceneParams->lights[3].color      = vec3(0.92f, 0.5f, 0.7f);
+            pPBRSceneParams->lights[3].intensity  = 0.5f;
             pPBRSceneParams->iblNumEnvLevels      = envNumLevels[gIBLIndex];
+            pPBRSceneParams->iblIndex             = gIBLIndex;
+            pPBRSceneParams->iblDiffuseStrength   = gIBLDiffuseStrength;
+            pPBRSceneParams->iblSpecularStrength  = gIBLSpecularStrength;
 
             // Draw environment
             {
@@ -739,8 +916,12 @@ int main(int argc, char** argv)
 
                 // SceneParmas (b0)
                 mat4 mvp = projMat * viewMat * moveUp;
-                vkCmdPushConstants(cmdBuf.CommandBuffer, envPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &mvp);
-                vkCmdPushConstants(cmdBuf.CommandBuffer, envPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &pPBRSceneParams->iblIndex);
+
+                EnvSceneParameters envSceneParams = {};
+                envSceneParams.IBLIndex           = gIBLIndex;
+                envSceneParams.MVP                = mvp;
+
+                vkCmdPushConstants(cmdBuf.CommandBuffer, envPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(EnvSceneParameters), &envSceneParams);
 
                 // Bind the Index Buffer
                 vkCmdBindIndexBuffer(cmdBuf.CommandBuffer, envGeoBuffers.indexBuffer.Buffer, 0, VK_INDEX_TYPE_UINT32);
@@ -753,7 +934,7 @@ int main(int argc, char** argv)
                 vkCmdDrawIndexed(cmdBuf.CommandBuffer, envGeoBuffers.numIndices, 1, 0, 0, 0);
             }
 
-            // Draw material models
+            // Draw sample spheres
             {
                 VkDescriptorBufferBindingInfoEXT descriptorBufferBindingInfo = {VK_STRUCTURE_TYPE_DESCRIPTOR_BUFFER_BINDING_INFO_EXT};
                 descriptorBufferBindingInfo.pNext                            = nullptr;
@@ -773,251 +954,146 @@ int main(int argc, char** argv)
                     &descriptorBufferOffsets);
 
                 // Select which model to draw
-                const GeometryBuffers& geoBuffers = matGeoBuffers[gModelIndex];
+                const GeometryBuffers& geoBuffers    = matGeoBuffers[gModelIndex];
+                uint32_t               geoIndexCount = geoBuffers.numIndices;
 
-                // Index buffer
+                // Bind the Index Buffer
                 vkCmdBindIndexBuffer(cmdBuf.CommandBuffer, geoBuffers.indexBuffer.Buffer, 0, VK_INDEX_TYPE_UINT32);
 
-                // Vertex buffers
-                VkBuffer vertexBuffers[] = {
-                    geoBuffers.positionBuffer.Buffer,
-                    geoBuffers.texCoordBuffer.Buffer,
-                    geoBuffers.normalBuffer.Buffer,
-                    geoBuffers.tangentBuffer.Buffer,
-                    geoBuffers.bitangentBuffer.Buffer,
-                };
-                VkDeviceSize offsets[] = {0, 0, 0, 0, 0};
-                vkCmdBindVertexBuffers(cmdBuf.CommandBuffer, 0, 5, vertexBuffers, offsets);
+                // Bind the Vertex Buffer
+                VkBuffer     vertexBuffers[] = {geoBuffers.positionBuffer.Buffer, geoBuffers.normalBuffer.Buffer};
+                VkDeviceSize offsets[]       = {0, 0};
+                vkCmdBindVertexBuffers(cmdBuf.CommandBuffer, 0, 2, vertexBuffers, offsets);
 
                 // Pipeline state
                 vkCmdBindPipeline(cmdBuf.CommandBuffer, VK_PIPELINE_BIND_POINT_GRAPHICS, pbrPipelineState);
 
-                const float yPos             = 0.0f;
-                uint32_t    materialIndex    = 0;
-                uint32_t    invertNormalMapY = false; // Invert if sphere
+                const float yPos = 0.0f;
 
-                // Material 0
+                // Copper
                 {
-                    glm::mat4 modelMat = glm::translate(vec3(-4.5f, yPos, 4.5f));
+                    glm::mat4 modelMat      = glm::translate(vec3(-3, yPos, 3));
+                    uint32_t  materialIndex = 0;
 
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &modelMat);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &materialIndex);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4) + sizeof(uint32_t), sizeof(uint32_t), &invertNormalMapY);
-                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoBuffers.numIndices, 1, 0, 0, 0);
+                    DrawParameters drawParams = {};
+                    drawParams.ModelMatrix    = modelMat;
+                    drawParams.MaterialIndex  = materialIndex;
 
-                    if (materialIndex < (materialTexturesSets.size() - 1)) {
-                        ++materialIndex;
-                    }
+                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(DrawParameters), &drawParams);
+
+                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoIndexCount, 1, 0, 0, 0);
                 }
 
-                // Material 1
+                // Gold
                 {
-                    glm::mat4 modelMat = glm::translate(vec3(-1.5f, yPos, 4.5f));
+                    glm::mat4 modelMat      = glm::translate(vec3(0, yPos, 3));
+                    uint32_t  materialIndex = 1;
 
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &modelMat);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &materialIndex);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4) + sizeof(uint32_t), sizeof(uint32_t), &invertNormalMapY);
-                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoBuffers.numIndices, 1, 0, 0, 0);
+                    DrawParameters drawParams = {};
+                    drawParams.ModelMatrix    = modelMat;
+                    drawParams.MaterialIndex  = materialIndex;
 
-                    if (materialIndex < (materialTexturesSets.size() - 1)) {
-                        ++materialIndex;
-                    }
+                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(DrawParameters), &drawParams);
+
+                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoIndexCount, 1, 0, 0, 0);
                 }
 
-                // Material 2
+                // Silver
                 {
-                    glm::mat4 modelMat = glm::translate(vec3(1.5f, yPos, 4.5f));
+                    glm::mat4 modelMat      = glm::translate(vec3(3, yPos, 3));
+                    uint32_t  materialIndex = 2;
 
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &modelMat);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &materialIndex);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4) + sizeof(uint32_t), sizeof(uint32_t), &invertNormalMapY);
-                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoBuffers.numIndices, 1, 0, 0, 0);
+                    DrawParameters drawParams = {};
+                    drawParams.ModelMatrix    = modelMat;
+                    drawParams.MaterialIndex  = materialIndex;
 
-                    if (materialIndex < (materialTexturesSets.size() - 1)) {
-                        ++materialIndex;
-                    }
+                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(DrawParameters), &drawParams);
+
+                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoIndexCount, 1, 0, 0, 0);
                 }
 
-                // Material 3
+                // Zink
                 {
-                    glm::mat4 modelMat = glm::translate(vec3(4.5f, yPos, 4.5f));
+                    glm::mat4 modelMat      = glm::translate(vec3(-3, yPos, 0));
+                    uint32_t  materialIndex = 3;
 
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &modelMat);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &materialIndex);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4) + sizeof(uint32_t), sizeof(uint32_t), &invertNormalMapY);
-                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoBuffers.numIndices, 1, 0, 0, 0);
+                    DrawParameters drawParams = {};
+                    drawParams.ModelMatrix    = modelMat;
+                    drawParams.MaterialIndex  = materialIndex;
 
-                    if (materialIndex < (materialTexturesSets.size() - 1)) {
-                        ++materialIndex;
-                    }
+                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(DrawParameters), &drawParams);
+
+                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoIndexCount, 1, 0, 0, 0);
                 }
 
-                // Material 4
+                // Titanium
                 {
-                    glm::mat4 modelMat = glm::translate(vec3(-4.5f, yPos, 1.5f));
+                    glm::mat4 modelMat      = glm::translate(vec3(0, yPos, 0));
+                    uint32_t  materialIndex = 4;
 
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &modelMat);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &materialIndex);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4) + sizeof(uint32_t), sizeof(uint32_t), &invertNormalMapY);
-                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoBuffers.numIndices, 1, 0, 0, 0);
+                    DrawParameters drawParams = {};
+                    drawParams.ModelMatrix    = modelMat;
+                    drawParams.MaterialIndex  = materialIndex;
 
-                    if (materialIndex < (materialTexturesSets.size() - 1)) {
-                        ++materialIndex;
-                    }
+                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(DrawParameters), &drawParams);
+
+                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoIndexCount, 1, 0, 0, 0);
                 }
 
-                // Material 5
+                // Shiny Plastic
                 {
-                    glm::mat4 modelMat = glm::translate(vec3(-1.5f, yPos, 1.5f));
+                    glm::mat4 modelMat      = glm::translate(vec3(3, yPos, 0));
+                    uint32_t  materialIndex = 5;
 
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &modelMat);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &materialIndex);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4) + sizeof(uint32_t), sizeof(uint32_t), &invertNormalMapY);
-                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoBuffers.numIndices, 1, 0, 0, 0);
+                    DrawParameters drawParams = {};
+                    drawParams.ModelMatrix    = modelMat;
+                    drawParams.MaterialIndex  = materialIndex;
 
-                    if (materialIndex < (materialTexturesSets.size() - 1)) {
-                        ++materialIndex;
-                    }
+                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(DrawParameters), &drawParams);
+
+                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoIndexCount, 1, 0, 0, 0);
                 }
 
-                // Material 6
+                // Rough Plastic
                 {
-                    glm::mat4 modelMat = glm::translate(vec3(1.5f, yPos, 1.5f));
+                    glm::mat4 modelMat      = glm::translate(vec3(-3, yPos, -3));
+                    uint32_t  materialIndex = 6;
 
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &modelMat);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &materialIndex);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4) + sizeof(uint32_t), sizeof(uint32_t), &invertNormalMapY);
-                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoBuffers.numIndices, 1, 0, 0, 0);
+                    DrawParameters drawParams = {};
+                    drawParams.ModelMatrix    = modelMat;
+                    drawParams.MaterialIndex  = materialIndex;
 
-                    if (materialIndex < (materialTexturesSets.size() - 1)) {
-                        ++materialIndex;
-                    }
+                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(DrawParameters), &drawParams);
+
+                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoIndexCount, 1, 0, 0, 0);
                 }
 
-                // Material 7
+                // Rougher Plastic
                 {
-                    glm::mat4 modelMat = glm::translate(vec3(4.5f, yPos, 1.5f));
+                    glm::mat4 modelMat      = glm::translate(vec3(0, yPos, -3));
+                    uint32_t  materialIndex = 7;
 
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &modelMat);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &materialIndex);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4) + sizeof(uint32_t), sizeof(uint32_t), &invertNormalMapY);
-                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoBuffers.numIndices, 1, 0, 0, 0);
+                    DrawParameters drawParams = {};
+                    drawParams.ModelMatrix    = modelMat;
+                    drawParams.MaterialIndex  = materialIndex;
 
-                    if (materialIndex < (materialTexturesSets.size() - 1)) {
-                        ++materialIndex;
-                    }
+                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(DrawParameters), &drawParams);
+
+                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoIndexCount, 1, 0, 0, 0);
                 }
 
-                // Material 8
+                // Roughest Plastic
                 {
-                    glm::mat4 modelMat = glm::translate(vec3(-4.5f, yPos, -1.5f));
+                    glm::mat4 modelMat      = glm::translate(vec3(3, yPos, -3));
+                    uint32_t  materialIndex = 8;
 
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &modelMat);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &materialIndex);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4) + sizeof(uint32_t), sizeof(uint32_t), &invertNormalMapY);
-                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoBuffers.numIndices, 1, 0, 0, 0);
+                    DrawParameters drawParams = {};
+                    drawParams.ModelMatrix    = modelMat;
+                    drawParams.MaterialIndex  = materialIndex;
 
-                    if (materialIndex < (materialTexturesSets.size() - 1)) {
-                        ++materialIndex;
-                    }
-                }
+                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(DrawParameters), &drawParams);
 
-                // Material 9
-                {
-                    glm::mat4 modelMat = glm::translate(vec3(-1.5f, yPos, -1.5f));
-
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &modelMat);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &materialIndex);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4) + sizeof(uint32_t), sizeof(uint32_t), &invertNormalMapY);
-                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoBuffers.numIndices, 1, 0, 0, 0);
-
-                    if (materialIndex < (materialTexturesSets.size() - 1)) {
-                        ++materialIndex;
-                    }
-                }
-
-                // Material 10
-                {
-                    glm::mat4 modelMat = glm::translate(vec3(1.5f, yPos, -1.5f));
-
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &modelMat);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &materialIndex);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4) + sizeof(uint32_t), sizeof(uint32_t), &invertNormalMapY);
-                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoBuffers.numIndices, 1, 0, 0, 0);
-
-                    if (materialIndex < (materialTexturesSets.size() - 1)) {
-                        ++materialIndex;
-                    }
-                }
-
-                // Material 11
-                {
-                    glm::mat4 modelMat = glm::translate(vec3(4.5f, yPos, -1.5f));
-
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &modelMat);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &materialIndex);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4) + sizeof(uint32_t), sizeof(uint32_t), &invertNormalMapY);
-                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoBuffers.numIndices, 1, 0, 0, 0);
-
-                    if (materialIndex < (materialTexturesSets.size() - 1)) {
-                        ++materialIndex;
-                    }
-                }
-
-                // Material 12
-                {
-                    glm::mat4 modelMat = glm::translate(vec3(-4.5f, yPos, -4.5f));
-
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &modelMat);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &materialIndex);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4) + sizeof(uint32_t), sizeof(uint32_t), &invertNormalMapY);
-                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoBuffers.numIndices, 1, 0, 0, 0);
-
-                    if (materialIndex < (materialTexturesSets.size() - 1)) {
-                        ++materialIndex;
-                    }
-                }
-
-                // Material 13
-                {
-                    glm::mat4 modelMat = glm::translate(vec3(-1.5f, yPos, -4.5f));
-
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &modelMat);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &materialIndex);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4) + sizeof(uint32_t), sizeof(uint32_t), &invertNormalMapY);
-                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoBuffers.numIndices, 1, 0, 0, 0);
-
-                    if (materialIndex < (materialTexturesSets.size() - 1)) {
-                        ++materialIndex;
-                    }
-                }
-
-                // Material 14
-                {
-                    glm::mat4 modelMat = glm::translate(vec3(1.5f, yPos, -4.5f));
-
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &modelMat);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &materialIndex);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4) + sizeof(uint32_t), sizeof(uint32_t), &invertNormalMapY);
-                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoBuffers.numIndices, 1, 0, 0, 0);
-
-                    if (materialIndex < (materialTexturesSets.size() - 1)) {
-                        ++materialIndex;
-                    }
-                }
-
-                // Material 15
-                {
-                    glm::mat4 modelMat = glm::translate(vec3(4.5f, yPos, -4.5f));
-
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, 0, sizeof(glm::mat4), &modelMat);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4), sizeof(uint32_t), &materialIndex);
-                    vkCmdPushConstants(cmdBuf.CommandBuffer, pbrPipelineLayout.PipelineLayout, VK_SHADER_STAGE_ALL_GRAPHICS, sizeof(glm::mat4) + sizeof(uint32_t), sizeof(uint32_t), &invertNormalMapY);
-                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoBuffers.numIndices, 1, 0, 0, 0);
-
-                    if (materialIndex < (materialTexturesSets.size() - 1)) {
-                        ++materialIndex;
-                    }
+                    vkCmdDrawIndexed(cmdBuf.CommandBuffer, geoIndexCount, 1, 0, 0, 0);
                 }
             }
 
@@ -1079,7 +1155,7 @@ void CreatePBRPipeline(VulkanRenderer* pRenderer, VulkanPipelineLayout* pLayout)
     // Descriptor set layout
     {
         std::vector<VkDescriptorSetLayoutBinding> bindings = {};
-        // ConstantBuffer<SceneParameters>      SceneParams                   : register(b0);
+        // ConstantBuffer<SceneParameters>      SceneParams        : register(b0);
         {
             VkDescriptorSetLayoutBinding binding = {};
             binding.binding                      = 0;
@@ -1088,11 +1164,7 @@ void CreatePBRPipeline(VulkanRenderer* pRenderer, VulkanPipelineLayout* pLayout)
             binding.stageFlags                   = VK_SHADER_STAGE_ALL_GRAPHICS;
             bindings.push_back(binding);
         }
-
-        // DEFINE_AS_PUSH_CONSTANT
-        // ConstantBuffer<DrawParameters>       DrawParams                    : register(b1);
-
-        // StructuredBuffer<MaterialParameters> MaterialParams                : register(t2);
+        // StructuredBuffer<MaterialParameters> MaterialParams     : register(t2);
         {
             VkDescriptorSetLayoutBinding binding = {};
             binding.binding                      = 2;
@@ -1101,25 +1173,34 @@ void CreatePBRPipeline(VulkanRenderer* pRenderer, VulkanPipelineLayout* pLayout)
             binding.stageFlags                   = VK_SHADER_STAGE_ALL_GRAPHICS;
             bindings.push_back(binding);
         }
-        // Texture2D                            IBLIntegrationLUT             : register(t3);
-        {
-            VkDescriptorSetLayoutBinding binding = {};
-            binding.binding                      = 3;
-            binding.descriptorType               = VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE;
-            binding.descriptorCount              = 1;
-            binding.stageFlags                   = VK_SHADER_STAGE_ALL_GRAPHICS;
-            bindings.push_back(binding);
-        }
-        // Texture2D                            IBLIntegrationMultiscatterLUT : register(t4);
+        // SamplerState                         ClampedSampler     : register(s4);
         {
             VkDescriptorSetLayoutBinding binding = {};
             binding.binding                      = 4;
+            binding.descriptorType               = VK_DESCRIPTOR_TYPE_SAMPLER;
+            binding.descriptorCount              = 1;
+            binding.stageFlags                   = VK_SHADER_STAGE_ALL_GRAPHICS;
+            bindings.push_back(binding);
+        }
+        // SamplerState                         UWrapSampler       : register(s5);
+        {
+            VkDescriptorSetLayoutBinding binding = {};
+            binding.binding                      = 5;
+            binding.descriptorType               = VK_DESCRIPTOR_TYPE_SAMPLER;
+            binding.descriptorCount              = 1;
+            binding.stageFlags                   = VK_SHADER_STAGE_ALL_GRAPHICS;
+            bindings.push_back(binding);
+        }
+        // Texture2D                            BRDFLUT            : register(t10);
+        {
+            VkDescriptorSetLayoutBinding binding = {};
+            binding.binding                      = 10;
             binding.descriptorType               = VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE;
             binding.descriptorCount              = 1;
             binding.stageFlags                   = VK_SHADER_STAGE_ALL_GRAPHICS;
             bindings.push_back(binding);
         }
-        // Texture2D                            IBLIrradianceMaps[32]         : register(t16);
+        // Texture2D                            IrradianceMap[32]  : register(t16);
         {
             VkDescriptorSetLayoutBinding binding = {};
             binding.binding                      = 16;
@@ -1128,57 +1209,12 @@ void CreatePBRPipeline(VulkanRenderer* pRenderer, VulkanPipelineLayout* pLayout)
             binding.stageFlags                   = VK_SHADER_STAGE_ALL_GRAPHICS;
             bindings.push_back(binding);
         }
-        // Texture2D                            IBLEnvironmentMaps[32]        : register(t48);
+        // Texture2D                            EnvironmentMap[32] : register(t48);
         {
             VkDescriptorSetLayoutBinding binding = {};
             binding.binding                      = 48;
             binding.descriptorType               = VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE;
             binding.descriptorCount              = 32;
-            binding.stageFlags                   = VK_SHADER_STAGE_ALL_GRAPHICS;
-            bindings.push_back(binding);
-        }
-        // SamplerState                         IBLIntegrationSampler         : register(s32);
-        {
-            VkDescriptorSetLayoutBinding binding = {};
-            binding.binding                      = 32;
-            binding.descriptorType               = VK_DESCRIPTOR_TYPE_SAMPLER;
-            binding.descriptorCount              = 1;
-            binding.stageFlags                   = VK_SHADER_STAGE_ALL_GRAPHICS;
-            bindings.push_back(binding);
-        }
-        // SamplerState                         IBLMapSampler                 : register(s33);
-        {
-            VkDescriptorSetLayoutBinding binding = {};
-            binding.binding                      = 33;
-            binding.descriptorType               = VK_DESCRIPTOR_TYPE_SAMPLER;
-            binding.descriptorCount              = 1;
-            binding.stageFlags                   = VK_SHADER_STAGE_ALL_GRAPHICS;
-            bindings.push_back(binding);
-        }
-        // Texture2D    MaterialTextures[TOTAL_MATERIAL_TEXTURES] : register(t100);
-        {
-            VkDescriptorSetLayoutBinding binding = {};
-            binding.binding                      = 100;
-            binding.descriptorType               = VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE;
-            binding.descriptorCount              = TOTAL_MATERIAL_TEXTURES;
-            binding.stageFlags                   = VK_SHADER_STAGE_ALL_GRAPHICS;
-            bindings.push_back(binding);
-        }
-        // SamplerState MaterialSampler                           : register(s34);
-        {
-            VkDescriptorSetLayoutBinding binding = {};
-            binding.binding                      = 34;
-            binding.descriptorType               = VK_DESCRIPTOR_TYPE_SAMPLER;
-            binding.descriptorCount              = 1;
-            binding.stageFlags                   = VK_SHADER_STAGE_ALL_GRAPHICS;
-            bindings.push_back(binding);
-        }
-        // SamplerState MaterialNormalMapSampler                  : register(s35);
-        {
-            VkDescriptorSetLayoutBinding binding = {};
-            binding.binding                      = 35;
-            binding.descriptorType               = VK_DESCRIPTOR_TYPE_SAMPLER;
-            binding.descriptorCount              = 1;
             binding.stageFlags                   = VK_SHADER_STAGE_ALL_GRAPHICS;
             bindings.push_back(binding);
         }
@@ -1218,10 +1254,6 @@ void CreateEnvironmentPipeline(VulkanRenderer* pRenderer, VulkanPipelineLayout* 
     // Descriptor set layout
     {
         std::vector<VkDescriptorSetLayoutBinding> bindings = {};
-        // DEFINE_AS_PUSH_CONSTANT
-        // ConstantBuffer<SceneParameters> SceneParmas  : register(b0);
-
-        // SamplerState                    Sampler0     : register(s1);
         {
             VkDescriptorSetLayoutBinding binding = {};
             binding.binding                      = 1;
@@ -1230,7 +1262,6 @@ void CreateEnvironmentPipeline(VulkanRenderer* pRenderer, VulkanPipelineLayout* 
             binding.stageFlags                   = VK_SHADER_STAGE_ALL_GRAPHICS;
             bindings.push_back(binding);
         }
-        // Texture2D                       Textures[16] : register(t32);
         {
             VkDescriptorSetLayoutBinding binding = {};
             binding.binding                      = 32;
@@ -1308,9 +1339,7 @@ void CreateMaterialModels(
 {
     // Sphere
     {
-        TriMesh::Options options = {.enableTexCoords = true, .enableNormals = true, .enableTangents = true};
-
-        TriMesh mesh = TriMesh::Sphere(1, 256, 256, options);
+        TriMesh mesh = TriMesh::Sphere(1, 256, 256, {.enableNormals = true});
 
         GeometryBuffers buffers = {};
 
@@ -1336,15 +1365,6 @@ void CreateMaterialModels(
 
         CHECK_CALL(CreateBuffer(
             pRenderer,
-            SizeInBytes(mesh.GetTexCoords()),
-            DataPtr(mesh.GetTexCoords()),
-            VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
-            VMA_MEMORY_USAGE_GPU_ONLY,
-            0,
-            &buffers.texCoordBuffer));
-
-        CHECK_CALL(CreateBuffer(
-            pRenderer,
             SizeInBytes(mesh.GetNormals()),
             DataPtr(mesh.GetNormals()),
             VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
@@ -1352,34 +1372,13 @@ void CreateMaterialModels(
             0,
             &buffers.normalBuffer));
 
-        CHECK_CALL(CreateBuffer(
-            pRenderer,
-            SizeInBytes(mesh.GetTangents()),
-            DataPtr(mesh.GetTangents()),
-            VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
-            VMA_MEMORY_USAGE_GPU_ONLY,
-            0,
-            &buffers.tangentBuffer));
-
-        CHECK_CALL(CreateBuffer(
-            pRenderer,
-            SizeInBytes(mesh.GetBitangents()),
-            DataPtr(mesh.GetBitangents()),
-            VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
-            VMA_MEMORY_USAGE_GPU_ONLY,
-            0,
-            &buffers.bitangentBuffer));
-
         outGeomtryBuffers.push_back(buffers);
     }
 
     // Knob
     {
         TriMesh::Options options = {};
-        options.enableTexCoords  = true;
         options.enableNormals    = true;
-        options.enableTangents   = true;
-        options.invertTexCoordsV = true;
         options.applyTransform   = true;
         options.transformRotate  = glm::vec3(0, glm::radians(180.0f), 0);
 
@@ -1413,15 +1412,6 @@ void CreateMaterialModels(
 
         CHECK_CALL(CreateBuffer(
             pRenderer,
-            SizeInBytes(mesh.GetTexCoords()),
-            DataPtr(mesh.GetTexCoords()),
-            VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
-            VMA_MEMORY_USAGE_GPU_ONLY,
-            0,
-            &buffers.texCoordBuffer));
-
-        CHECK_CALL(CreateBuffer(
-            pRenderer,
             SizeInBytes(mesh.GetNormals()),
             DataPtr(mesh.GetNormals()),
             VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
@@ -1429,33 +1419,13 @@ void CreateMaterialModels(
             0,
             &buffers.normalBuffer));
 
-        CHECK_CALL(CreateBuffer(
-            pRenderer,
-            SizeInBytes(mesh.GetTangents()),
-            DataPtr(mesh.GetTangents()),
-            VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
-            VMA_MEMORY_USAGE_GPU_ONLY,
-            0,
-            &buffers.tangentBuffer));
-
-        CHECK_CALL(CreateBuffer(
-            pRenderer,
-            SizeInBytes(mesh.GetBitangents()),
-            DataPtr(mesh.GetBitangents()),
-            VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
-            VMA_MEMORY_USAGE_GPU_ONLY,
-            0,
-            &buffers.bitangentBuffer));
-
         outGeomtryBuffers.push_back(buffers);
     }
 
     // Monkey
     {
         TriMesh::Options options = {};
-        options.enableTexCoords  = true;
         options.enableNormals    = true;
-        options.enableTangents   = true;
         options.applyTransform   = true;
         options.transformRotate  = glm::vec3(0, glm::radians(180.0f), 0);
 
@@ -1489,15 +1459,6 @@ void CreateMaterialModels(
 
         CHECK_CALL(CreateBuffer(
             pRenderer,
-            SizeInBytes(mesh.GetTexCoords()),
-            DataPtr(mesh.GetTexCoords()),
-            VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
-            VMA_MEMORY_USAGE_GPU_ONLY,
-            0,
-            &buffers.texCoordBuffer));
-
-        CHECK_CALL(CreateBuffer(
-            pRenderer,
             SizeInBytes(mesh.GetNormals()),
             DataPtr(mesh.GetNormals()),
             VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
@@ -1505,30 +1466,21 @@ void CreateMaterialModels(
             0,
             &buffers.normalBuffer));
 
-        CHECK_CALL(CreateBuffer(
-            pRenderer,
-            SizeInBytes(mesh.GetTangents()),
-            DataPtr(mesh.GetTangents()),
-            VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
-            VMA_MEMORY_USAGE_GPU_ONLY,
-            0,
-            &buffers.tangentBuffer));
-
-        CHECK_CALL(CreateBuffer(
-            pRenderer,
-            SizeInBytes(mesh.GetBitangents()),
-            DataPtr(mesh.GetBitangents()),
-            VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
-            VMA_MEMORY_USAGE_GPU_ONLY,
-            0,
-            &buffers.bitangentBuffer));
-
         outGeomtryBuffers.push_back(buffers);
     }
 
-    // Cube
+    // Teapot
     {
-        TriMesh mesh = TriMesh::Cube(vec3(2), false, {.enableTexCoords = true, .enableNormals = true, .enableTangents = true});
+        TriMesh::Options options = {};
+        options.enableNormals    = true;
+        options.applyTransform   = true;
+        options.transformRotate  = glm::vec3(0, glm::radians(135.0f), 0);
+
+        TriMesh mesh;
+        if (!TriMesh::LoadOBJ(GetAssetPath("models/teapot.obj").string(), "", options, &mesh)) {
+            return;
+        }
+        mesh.ScaleToFit(2.0f);
 
         GeometryBuffers buffers = {};
 
@@ -1554,39 +1506,12 @@ void CreateMaterialModels(
 
         CHECK_CALL(CreateBuffer(
             pRenderer,
-            SizeInBytes(mesh.GetTexCoords()),
-            DataPtr(mesh.GetTexCoords()),
-            VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
-            VMA_MEMORY_USAGE_GPU_ONLY,
-            0,
-            &buffers.texCoordBuffer));
-
-        CHECK_CALL(CreateBuffer(
-            pRenderer,
             SizeInBytes(mesh.GetNormals()),
             DataPtr(mesh.GetNormals()),
             VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
             VMA_MEMORY_USAGE_GPU_ONLY,
             0,
             &buffers.normalBuffer));
-
-        CHECK_CALL(CreateBuffer(
-            pRenderer,
-            SizeInBytes(mesh.GetTangents()),
-            DataPtr(mesh.GetTangents()),
-            VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
-            VMA_MEMORY_USAGE_GPU_ONLY,
-            0,
-            &buffers.tangentBuffer));
-
-        CHECK_CALL(CreateBuffer(
-            pRenderer,
-            SizeInBytes(mesh.GetBitangents()),
-            DataPtr(mesh.GetBitangents()),
-            VK_BUFFER_USAGE_VERTEX_BUFFER_BIT,
-            VMA_MEMORY_USAGE_GPU_ONLY,
-            0,
-            &buffers.bitangentBuffer));
 
         outGeomtryBuffers.push_back(buffers);
     }
@@ -1595,7 +1520,6 @@ void CreateMaterialModels(
 void CreateIBLTextures(
     VulkanRenderer*           pRenderer,
     VulkanImage*              pBRDFLUT,
-    VulkanImage*              pMultiscatterBRDFLUT,
     std::vector<VulkanImage>& outIrradianceTextures,
     std::vector<VulkanImage>& outEnvironmentTextures,
     std::vector<uint32_t>&    outEnvNumLevels)
@@ -1618,24 +1542,6 @@ void CreateIBLTextures(
             pBRDFLUT));
     }
 
-    // Multiscatter BRDF LUT
-    {
-        auto bitmap = LoadImage32f(GetAssetPath("IBL/brdf_lut_ms.hdr"));
-        if (bitmap.Empty()) {
-            assert(false && "Load image failed");
-            return;
-        }
-
-        CHECK_CALL(CreateTexture(
-            pRenderer,
-            bitmap.GetWidth(),
-            bitmap.GetHeight(),
-            VK_FORMAT_R32G32B32A32_SFLOAT,
-            bitmap.GetSizeInBytes(),
-            bitmap.GetPixels(),
-            pMultiscatterBRDFLUT));
-    }
-
     auto                               iblDir = GetAssetPath("IBL");
     std::vector<std::filesystem::path> iblFiles;
     for (auto& entry : std::filesystem::directory_iterator(iblDir)) {
@@ -1652,12 +1558,11 @@ void CreateIBLTextures(
 
     size_t maxEntries = std::min<size_t>(gMaxIBLs, iblFiles.size());
     for (size_t i = 0; i < maxEntries; ++i) {
-        std::filesystem::path iblFile = iblFiles[i];
+        auto& iblFile = iblFiles[i];
 
         IBLMaps ibl = {};
         if (!LoadIBLMaps32f(iblFile, &ibl)) {
             GREX_LOG_ERROR("failed to load: " << iblFile);
-            assert(false && "IBL maps load failed failed");
             return;
         }
 
@@ -1665,7 +1570,7 @@ void CreateIBLTextures(
 
         // Irradiance
         {
-            VulkanImage texture = {};
+            VulkanImage texture;
             CHECK_CALL(CreateTexture(
                 pRenderer,
                 ibl.irradianceMap.GetWidth(),
@@ -1698,7 +1603,7 @@ void CreateIBLTextures(
                 levelHeight >>= 1;
             }
 
-            VulkanImage texture = {};
+            VulkanImage texture;
             CHECK_CALL(CreateTexture(
                 pRenderer,
                 ibl.baseWidth,
@@ -1714,134 +1619,6 @@ void CreateIBLTextures(
         gIBLNames.push_back(iblFile.filename().replace_extension().string());
 
         GREX_LOG_INFO("Loaded " << iblFile);
-    }
-}
-
-void CreateMaterials(
-    VulkanRenderer*                  pRenderer,
-    MaterialTextures&                outDefaultMaterialTextures,
-    std::vector<MaterialTextures>&   outMaterialTexturesSets,
-    std::vector<MaterialParameters>& outMaterialParametersSets)
-{
-    // Default material textures
-    {
-        PixelRGBA8u purplePixel = {0, 0, 0, 255};
-        PixelRGBA8u blackPixel  = {0, 0, 0, 255};
-        PixelRGBA8u whitePixel  = {255, 255, 255, 255};
-
-        CHECK_CALL(CreateTexture(pRenderer, 1, 1, VK_FORMAT_R8G8B8A8_UNORM, sizeof(PixelRGBA8u), &purplePixel, &outDefaultMaterialTextures.baseColorTexture));
-        CHECK_CALL(CreateTexture(pRenderer, 1, 1, VK_FORMAT_R8G8B8A8_UNORM, sizeof(PixelRGBA8u), &blackPixel, &outDefaultMaterialTextures.normalTexture));
-        CHECK_CALL(CreateTexture(pRenderer, 1, 1, VK_FORMAT_R8G8B8A8_UNORM, sizeof(PixelRGBA8u), &blackPixel, &outDefaultMaterialTextures.roughnessTexture));
-        CHECK_CALL(CreateTexture(pRenderer, 1, 1, VK_FORMAT_R8G8B8A8_UNORM, sizeof(PixelRGBA8u), &blackPixel, &outDefaultMaterialTextures.metallicTexture));
-    }
-
-    // Texture directory
-    auto texturesDir = GetAssetPath("textures");
-
-    // Material files - limit to 16 since there's 16 objects draws
-    std::vector<std::filesystem::path> materialFiles = {
-        texturesDir / "bark_brown_02" / "material.mat",
-        texturesDir / "bark_willow" / "material.mat",
-        texturesDir / "brick_4" / "material.mat",
-        texturesDir / "castle_brick_02_red" / "material.mat",
-        texturesDir / "dark_brick_wall" / "material.mat",
-        texturesDir / "factory_wall" / "material.mat",
-        texturesDir / "green_metal_rust" / "material.mat",
-        texturesDir / "hexagonal_concrete_paving" / "material.mat",
-        texturesDir / "metal_grate_rusty" / "material.mat",
-        texturesDir / "metal_plate" / "material.mat",
-        texturesDir / "mud_cracked_dry_riverbed_002" / "material.mat",
-        texturesDir / "pavement_02" / "material.mat",
-        texturesDir / "rough_plaster_broken" / "material.mat",
-        texturesDir / "rusty_metal_02" / "material.mat",
-        texturesDir / "weathered_planks" / "material.mat",
-        texturesDir / "wood_table_001" / "material.mat",
-    };
-
-    size_t maxEntries = materialFiles.size();
-    for (size_t i = 0; i < maxEntries; ++i) {
-        auto materialFile = materialFiles[i];
-
-        std::ifstream is = std::ifstream(materialFile.string().c_str());
-        if (!is.is_open()) {
-            assert(false && "faild to open material file");
-        }
-
-        MaterialTextures   materialTextures = outDefaultMaterialTextures;
-        MaterialParameters materialParams   = {};
-
-        while (!is.eof()) {
-            VulkanImage*          pTargetTexture = {};
-            std::filesystem::path textureFile   = "";
-
-            std::string key;
-            is >> key;
-            if (key == "basecolor") {
-                is >> textureFile;
-                pTargetTexture = &materialTextures.baseColorTexture;
-            }
-            else if (key == "normal") {
-                is >> textureFile;
-                pTargetTexture = &materialTextures.normalTexture;
-            }
-            else if (key == "roughness") {
-                is >> textureFile;
-                pTargetTexture = &materialTextures.roughnessTexture;
-            }
-            else if (key == "metallic") {
-                is >> textureFile;
-                pTargetTexture = &materialTextures.metallicTexture;
-            }
-            else if (key == "specular") {
-                is >> materialParams.specular;
-            }
-
-            if (textureFile.empty()) {
-                continue;
-            }
-
-            auto cwd    = materialFile.parent_path().filename();
-            textureFile = "textures" / cwd / textureFile;
-
-            auto bitmap = LoadImage8u(textureFile);
-            if (!bitmap.Empty()) {
-                MipmapRGBA8u mipmap = MipmapRGBA8u(
-                    bitmap,
-                    BITMAP_SAMPLE_MODE_WRAP,
-                    BITMAP_SAMPLE_MODE_WRAP,
-                    BITMAP_FILTER_MODE_NEAREST);
-
-                std::vector<MipOffset> mipOffsets;
-                for (auto& srcOffset : mipmap.GetOffsets()) {
-                    MipOffset dstOffset = {};
-                    dstOffset.Offset    = srcOffset;
-                    dstOffset.RowStride = mipmap.GetRowStride();
-                    mipOffsets.push_back(dstOffset);
-                }
-
-                CHECK_CALL(CreateTexture(
-                    pRenderer,
-                    mipmap.GetWidth(0),
-                    mipmap.GetHeight(0),
-                    VK_FORMAT_R8G8B8A8_UNORM,
-                    mipOffsets,
-                    mipmap.GetSizeInBytes(),
-                    mipmap.GetPixels(),
-                    &(*pTargetTexture)));
-
-                GREX_LOG_INFO("Created texture from " << textureFile);
-            }
-            else {
-                GREX_LOG_ERROR("Failed to load: " << textureFile);
-                assert(false && "Failed to load texture!");
-            }
-        }
-
-        outMaterialTexturesSets.push_back(materialTextures);
-        outMaterialParametersSets.push_back(materialParams);
-
-        // Use directory name for material name
-        gMaterialNames.push_back(materialFile.parent_path().filename().string());
     }
 }
 
@@ -1867,16 +1644,14 @@ void CreateDescriptorBuffer(
 }
 
 void WritePBRDescriptors(
-    VulkanRenderer*                pRenderer,
-    VkDescriptorSetLayout          descriptorSetLayout,
-    VulkanBuffer*                  pDescriptorBuffer,
-    VulkanBuffer*                  pSceneParamsBuffer,
-    VulkanBuffer*                  pMaterialBuffer,
-    std::vector<MaterialTextures>& materialTextureSets,
-    const VulkanImage*             pBRDFLUT,
-    const VulkanImage*             pMultiscatterBRDFLUT,
-    std::vector<VulkanImage>&      irrTextures,
-    std::vector<VulkanImage>&      envTextures)
+    VulkanRenderer*           pRenderer,
+    VkDescriptorSetLayout     descriptorSetLayout,
+    VulkanBuffer*             pDescriptorBuffer,
+    const VulkanBuffer*       pSceneParamsBuffer,
+    const VulkanBuffer*       pMaterialsBuffer,
+    const VulkanImage*        pBRDFLUT,
+    std::vector<VulkanImage>& irradianceTextures,
+    std::vector<VulkanImage>& envTextures)
 {
     char* pDescriptorBufferStartAddress = nullptr;
     CHECK_CALL(vmaMapMemory(
@@ -1884,7 +1659,7 @@ void WritePBRDescriptors(
         pDescriptorBuffer->Allocation,
         reinterpret_cast<void**>(&pDescriptorBufferStartAddress)));
 
-   // ConstantBuffer<SceneParameters>      SceneParams                                : register(b0);
+    // ConstantBuffer<SceneParameters>    SceneParams           : register(b0);
     WriteDescriptor(
         pRenderer,
         pDescriptorBufferStartAddress,
@@ -1894,10 +1669,10 @@ void WritePBRDescriptors(
         VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER,
         pSceneParamsBuffer);
 
-    // DEFINE_AS_PUSH_CONSTANT
-    // ConstantBuffer<DrawParameters>       DrawParams                                 : register(b1);
+    // Set via push constants
+    // ConstantBuffer<DrawParameters>     DrawParams            : register(b1);
 
-    // StructuredBuffer<MaterialParameters> MaterialParams                             : register(t2);
+    // ConstantBuffer<MaterialParameters> MaterialParams        : register(b2);
     WriteDescriptor(
         pRenderer,
         pDescriptorBufferStartAddress,
@@ -1905,107 +1680,10 @@ void WritePBRDescriptors(
         2, // binding
         0, // arrayElement
         VK_DESCRIPTOR_TYPE_STORAGE_BUFFER,
-        pMaterialBuffer);
+        pMaterialsBuffer);
 
-    // Texture2D                            IBLIntegrationLUT                          : register(t3);
-    {
-        VkImageView imageView = VK_NULL_HANDLE;
-        CHECK_CALL(CreateImageView(
-            pRenderer,
-            pBRDFLUT,
-            VK_IMAGE_VIEW_TYPE_2D,
-            VK_FORMAT_R32G32B32A32_SFLOAT,
-            GREX_ALL_SUBRESOURCES,
-            &imageView));
-
-        WriteDescriptor(
-            pRenderer,
-            pDescriptorBufferStartAddress,
-            descriptorSetLayout,
-            3, // binding
-            0, // arrayElement
-            VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE,
-            imageView,
-            VK_IMAGE_LAYOUT_GENERAL);
-    }
-
-    // Texture2D                            IBLIntegrationMultiscatterLUT              : register(t4);
-    {
-        VkImageView imageView = VK_NULL_HANDLE;
-        CHECK_CALL(CreateImageView(
-            pRenderer,
-            pMultiscatterBRDFLUT,
-            VK_IMAGE_VIEW_TYPE_2D,
-            VK_FORMAT_R32G32B32A32_SFLOAT,
-            GREX_ALL_SUBRESOURCES,
-            &imageView));
-
-        WriteDescriptor(
-            pRenderer,
-            pDescriptorBufferStartAddress,
-            descriptorSetLayout,
-            4, // binding
-            0, // arrayElement
-            VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE,
-            imageView,
-            VK_IMAGE_LAYOUT_GENERAL);
-    }
-
-    // Texture2D                            IBLIrradianceMaps[32]                      : register(t16);
-    {
-        uint32_t arrayIndex = 0;
-        for (auto& image : irrTextures) {
-            VkImageView imageView = VK_NULL_HANDLE;
-            CHECK_CALL(CreateImageView(
-                pRenderer,
-                &image,
-                VK_IMAGE_VIEW_TYPE_2D,
-                VK_FORMAT_R32G32B32A32_SFLOAT,
-                GREX_ALL_SUBRESOURCES,
-                &imageView));
-
-            WriteDescriptor(
-                pRenderer,
-                pDescriptorBufferStartAddress,
-                descriptorSetLayout,
-                16,         // binding
-                arrayIndex, // arrayElement
-                VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE,
-                imageView,
-                VK_IMAGE_LAYOUT_GENERAL);
-
-            arrayIndex++;
-        }
-    }
-
-    // Texture2D                            IBLEnvironmentMaps[32]                     : register(t48);
-    {
-        uint32_t arrayIndex = 0;
-        for (auto& image : envTextures) {
-            VkImageView imageView = VK_NULL_HANDLE;
-            CHECK_CALL(CreateImageView(
-                pRenderer,
-                &image,
-                VK_IMAGE_VIEW_TYPE_2D,
-                VK_FORMAT_R32G32B32A32_SFLOAT,
-                GREX_ALL_SUBRESOURCES,
-                &imageView));
-
-            WriteDescriptor(
-                pRenderer,
-                pDescriptorBufferStartAddress,
-                descriptorSetLayout,
-                48,         // binding
-                arrayIndex, // arrayElement
-                VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE,
-                imageView,
-                VK_IMAGE_LAYOUT_GENERAL);
-
-            arrayIndex++;
-        }
-    }
-
-    // SamplerState                         IBLIntegrationSampler                      : register(s32);
+    // Samplers are setup in the immutable samplers in the DescriptorSetLayout
+    // SamplerState                       IBLIntegrationSampler : register(s4);
     {
         VkSamplerCreateInfo samplerInfo     = {VK_STRUCTURE_TYPE_SAMPLER_CREATE_INFO};
         samplerInfo.flags                   = 0;
@@ -2025,23 +1703,23 @@ void WritePBRDescriptors(
         samplerInfo.borderColor             = VK_BORDER_COLOR_FLOAT_OPAQUE_BLACK;
         samplerInfo.unnormalizedCoordinates = VK_FALSE;
 
-        VkSampler iblIntegrationSampler = VK_NULL_HANDLE;
+        VkSampler clampedSampler = VK_NULL_HANDLE;
         CHECK_CALL(vkCreateSampler(
             pRenderer->Device,
             &samplerInfo,
             nullptr,
-            &iblIntegrationSampler));
+            &clampedSampler));
 
         WriteDescriptor(
             pRenderer,
             pDescriptorBufferStartAddress,
             descriptorSetLayout,
-            32, // binding
-            0,  // arrayElement
-            iblIntegrationSampler);
+            4, // binding
+            0, // arrayElement
+            clampedSampler);
     }
 
-    // SamplerState                         IBLMapSampler                              : register(s33);
+    // SamplerState                         UWrapSampler       : register(s5);
     {
         VkSamplerCreateInfo samplerInfo     = {VK_STRUCTURE_TYPE_SAMPLER_CREATE_INFO};
         samplerInfo.flags                   = 0;
@@ -2061,137 +1739,106 @@ void WritePBRDescriptors(
         samplerInfo.borderColor             = VK_BORDER_COLOR_FLOAT_OPAQUE_BLACK;
         samplerInfo.unnormalizedCoordinates = VK_FALSE;
 
-        VkSampler iblMapSampler = VK_NULL_HANDLE;
+        VkSampler uWrapSampler = VK_NULL_HANDLE;
         CHECK_CALL(vkCreateSampler(
             pRenderer->Device,
             &samplerInfo,
             nullptr,
-            &iblMapSampler));
+            &uWrapSampler));
 
         WriteDescriptor(
             pRenderer,
             pDescriptorBufferStartAddress,
             descriptorSetLayout,
-            33, // binding
-            0,  // arrayElement
-            iblMapSampler);
+            5, // binding
+            0, // arrayElement
+            uWrapSampler);
     }
 
-    // Texture2D                            MaterialTextures[TOTAL_MATERIAL_TEXTURES]  : register(t100);
+    // Texture2D                            BRDFLUT            : register(t10);
     {
-        uint32_t arrayIndex = 0;
-        for (auto& materialTextures : materialTextureSets) {
-            VulkanImage* textureImages[] = {
-                &materialTextures.baseColorTexture,
-                &materialTextures.normalTexture,
-                &materialTextures.roughnessTexture,
-                &materialTextures.metallicTexture};
+        VkImageView imageView = VK_NULL_HANDLE;
+        CHECK_CALL(CreateImageView(
+            pRenderer,
+            pBRDFLUT,
+            VK_IMAGE_VIEW_TYPE_2D,
+            VK_FORMAT_R32G32B32A32_SFLOAT,
+            GREX_ALL_SUBRESOURCES,
+            &imageView));
 
-            for (auto& image : textureImages) {
-                VkImageView imageView = VK_NULL_HANDLE;
-                CHECK_CALL(CreateImageView(
-                    pRenderer,
-                    image,
-                    VK_IMAGE_VIEW_TYPE_2D,
-                    VK_FORMAT_R8G8B8A8_UNORM,
-                    GREX_ALL_SUBRESOURCES,
-                    &imageView));
+        WriteDescriptor(
+            pRenderer,
+            pDescriptorBufferStartAddress,
+            descriptorSetLayout,
+            10, // binding
+            0,  // arrayElement
+            VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE,
+            imageView,
+            VK_IMAGE_LAYOUT_GENERAL);
+    }
 
-                WriteDescriptor(
-                    pRenderer,
-                    pDescriptorBufferStartAddress,
-                    descriptorSetLayout,
-                    100,        // binding
-                    arrayIndex, // arrayElement
-                    VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE,
-                    imageView,
-                    VK_IMAGE_LAYOUT_GENERAL);
+    // Texture2D                            IrradianceMap[32]  : register(t16);
+    {
+        uint32_t arrayElement = 0;
+        for (auto& irradianceTexture : irradianceTextures) {
+            VkImageView imageView = VK_NULL_HANDLE;
+            CHECK_CALL(CreateImageView(
+                pRenderer,
+                &irradianceTexture,
+                VK_IMAGE_VIEW_TYPE_2D,
+                VK_FORMAT_R32G32B32A32_SFLOAT,
+                GREX_ALL_SUBRESOURCES,
+                &imageView));
 
-                arrayIndex++;
-            }
+            WriteDescriptor(
+                pRenderer,
+                pDescriptorBufferStartAddress,
+                descriptorSetLayout,
+                16,           // binding
+                arrayElement, // arrayElement
+                VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE,
+                imageView,
+                VK_IMAGE_LAYOUT_GENERAL);
+
+            arrayElement++;
         }
     }
 
-    // SamplerState                         MaterialSampler                            : register(s34);
+    // Texture2D                            EnvironmentMap[32] : register(t48);
     {
-        VkSamplerCreateInfo samplerInfo     = {VK_STRUCTURE_TYPE_SAMPLER_CREATE_INFO};
-        samplerInfo.flags                   = 0;
-        samplerInfo.magFilter               = VK_FILTER_LINEAR;
-        samplerInfo.minFilter               = VK_FILTER_LINEAR;
-        samplerInfo.mipmapMode              = VK_SAMPLER_MIPMAP_MODE_LINEAR;
-        samplerInfo.addressModeU            = VK_SAMPLER_ADDRESS_MODE_REPEAT;
-        samplerInfo.addressModeV            = VK_SAMPLER_ADDRESS_MODE_REPEAT;
-        samplerInfo.addressModeW            = VK_SAMPLER_ADDRESS_MODE_CLAMP_TO_EDGE;
-        samplerInfo.mipLodBias              = 0;
-        samplerInfo.anisotropyEnable        = VK_FALSE;
-        samplerInfo.maxAnisotropy           = 0;
-        samplerInfo.compareEnable           = VK_TRUE;
-        samplerInfo.compareOp               = VK_COMPARE_OP_LESS_OR_EQUAL;
-        samplerInfo.minLod                  = 0;
-        samplerInfo.maxLod                  = FLT_MAX;
-        samplerInfo.borderColor             = VK_BORDER_COLOR_FLOAT_OPAQUE_BLACK;
-        samplerInfo.unnormalizedCoordinates = VK_FALSE;
+        uint32_t arrayElement = 0;
+        for (auto& envTexture : envTextures) {
+            VkImageView imageView = VK_NULL_HANDLE;
+            CHECK_CALL(CreateImageView(
+                pRenderer,
+                &envTexture,
+                VK_IMAGE_VIEW_TYPE_2D,
+                VK_FORMAT_R32G32B32A32_SFLOAT,
+                GREX_ALL_SUBRESOURCES,
+                &imageView));
 
-        VkSampler materialSampler = VK_NULL_HANDLE;
-        CHECK_CALL(vkCreateSampler(
-            pRenderer->Device,
-            &samplerInfo,
-            nullptr,
-            &materialSampler));
+            WriteDescriptor(
+                pRenderer,
+                pDescriptorBufferStartAddress,
+                descriptorSetLayout,
+                48,           // binding
+                arrayElement, // arrayElement
+                VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE,
+                imageView,
+                VK_IMAGE_LAYOUT_GENERAL);
 
-        WriteDescriptor(
-            pRenderer,
-            pDescriptorBufferStartAddress,
-            descriptorSetLayout,
-            34, // binding
-            0,  // arrayElement
-            materialSampler);
-    }
-
-    // SamplerState                         MaterialNormalMapSampler                   : register(s35);
-    {
-        VkSamplerCreateInfo samplerInfo     = {VK_STRUCTURE_TYPE_SAMPLER_CREATE_INFO};
-        samplerInfo.flags                   = 0;
-        samplerInfo.magFilter               = VK_FILTER_LINEAR;
-        samplerInfo.minFilter               = VK_FILTER_LINEAR;
-        samplerInfo.mipmapMode              = VK_SAMPLER_MIPMAP_MODE_LINEAR;
-        samplerInfo.addressModeU            = VK_SAMPLER_ADDRESS_MODE_REPEAT;
-        samplerInfo.addressModeV            = VK_SAMPLER_ADDRESS_MODE_CLAMP_TO_EDGE;
-        samplerInfo.addressModeW            = VK_SAMPLER_ADDRESS_MODE_CLAMP_TO_EDGE;
-        samplerInfo.mipLodBias              = 0;
-        samplerInfo.anisotropyEnable        = VK_FALSE;
-        samplerInfo.maxAnisotropy           = 0;
-        samplerInfo.compareEnable           = VK_TRUE;
-        samplerInfo.compareOp               = VK_COMPARE_OP_LESS_OR_EQUAL;
-        samplerInfo.minLod                  = 0;
-        samplerInfo.maxLod                  = FLT_MAX;
-        samplerInfo.borderColor             = VK_BORDER_COLOR_FLOAT_OPAQUE_BLACK;
-        samplerInfo.unnormalizedCoordinates = VK_FALSE;
-
-        VkSampler materialNormalSampler = VK_NULL_HANDLE;
-        CHECK_CALL(vkCreateSampler(
-            pRenderer->Device,
-            &samplerInfo,
-            nullptr,
-            &materialNormalSampler));
-
-        WriteDescriptor(
-            pRenderer,
-            pDescriptorBufferStartAddress,
-            descriptorSetLayout,
-            35, // binding
-            0, // arrayElement
-            materialNormalSampler);
+            arrayElement++;
+        }
     }
 
     vmaUnmapMemory(pRenderer->Allocator, pDescriptorBuffer->Allocation);
 }
 
 void WriteEnvDescriptors(
-    VulkanRenderer*          pRenderer,
-    VkDescriptorSetLayout    descriptorSetLayout,
-    VulkanBuffer*            pDescriptorBuffer,
-    std::vector<VulkanImage> envTextures)
+    VulkanRenderer*           pRenderer,
+    VkDescriptorSetLayout     descriptorSetLayout,
+    VulkanBuffer*             pDescriptorBuffer,
+    std::vector<VulkanImage>& envTextures)
 {
     char* pDescriptorBufferStartAddress = nullptr;
     CHECK_CALL(vmaMapMemory(
@@ -2199,10 +1846,10 @@ void WriteEnvDescriptors(
         pDescriptorBuffer->Allocation,
         reinterpret_cast<void**>(&pDescriptorBufferStartAddress)));
 
-    // DEFINE_AS_PUSH_CONSTANT
-    // ConstantBuffer<SceneParameters> SceneParmas  : register(b0);
+    // set via push constants
+    // ConstantBuffer<SceneParameters> SceneParams       : register(b0);
 
-    // SamplerState                    Sampler0     : register(s1);
+    // SamplerState                    IBLMapSampler     : register(s1);
     {
         VkSamplerCreateInfo samplerInfo     = {VK_STRUCTURE_TYPE_SAMPLER_CREATE_INFO};
         samplerInfo.flags                   = 0;
@@ -2238,14 +1885,14 @@ void WriteEnvDescriptors(
             uWrapSampler);
     }
 
-    // Texture2D                       Textures[16] : register(t32);
+    // Texture2D                       IBLEnvironmentMap : register(t2);
     {
-        uint32_t arrayIndex = 0;
-        for (auto& image : envTextures) {
+        uint32_t arrayElement = 0;
+        for (auto& envTexture : envTextures) {
             VkImageView imageView = VK_NULL_HANDLE;
             CHECK_CALL(CreateImageView(
                 pRenderer,
-                &image,
+                &envTexture,
                 VK_IMAGE_VIEW_TYPE_2D,
                 VK_FORMAT_R32G32B32A32_SFLOAT,
                 GREX_ALL_SUBRESOURCES,
@@ -2255,15 +1902,190 @@ void WriteEnvDescriptors(
                 pRenderer,
                 pDescriptorBufferStartAddress,
                 descriptorSetLayout,
-                32,         // binding
-                arrayIndex, // arrayElement
+                32,           // binding
+                arrayElement, // arrayElement
                 VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE,
                 imageView,
                 VK_IMAGE_LAYOUT_GENERAL);
 
-            arrayIndex++;
+            arrayElement++;
         }
     }
 
     vmaUnmapMemory(pRenderer->Allocator, pDescriptorBuffer->Allocation);
 }
+
+/*
+void CreatePBRRootSig(DxRenderer* pRenderer, ID3D12RootSignature** ppRootSig)
+{
+    // BRDFLUT (t10)
+    D3D12_DESCRIPTOR_RANGE range1            = {};
+    range1.RangeType                         = D3D12_DESCRIPTOR_RANGE_TYPE_SRV;
+    range1.NumDescriptors                    = 1;
+    range1.BaseShaderRegister                = 10;
+    range1.RegisterSpace                     = 0;
+    range1.OffsetInDescriptorsFromTableStart = D3D12_DESCRIPTOR_RANGE_OFFSET_APPEND;
+
+    // IrradianceMap (t16)
+    D3D12_DESCRIPTOR_RANGE range2            = {};
+    range2.RangeType                         = D3D12_DESCRIPTOR_RANGE_TYPE_SRV;
+    range2.NumDescriptors                    = gMaxIBLs;
+    range2.BaseShaderRegister                = 16;
+    range2.RegisterSpace                     = 0;
+    range2.OffsetInDescriptorsFromTableStart = D3D12_DESCRIPTOR_RANGE_OFFSET_APPEND;
+
+    // EnvironmentMap (t32)
+    D3D12_DESCRIPTOR_RANGE range3            = {};
+    range3.RangeType                         = D3D12_DESCRIPTOR_RANGE_TYPE_SRV;
+    range3.NumDescriptors                    = gMaxIBLs;
+    range3.BaseShaderRegister                = 48;
+    range3.RegisterSpace                     = 0;
+    range3.OffsetInDescriptorsFromTableStart = D3D12_DESCRIPTOR_RANGE_OFFSET_APPEND;
+
+    D3D12_ROOT_PARAMETER rootParameters[6] = {};
+    // SceneParams (b0)
+    rootParameters[0].ParameterType             = D3D12_ROOT_PARAMETER_TYPE_CBV;
+    rootParameters[0].Descriptor.ShaderRegister = 0;
+    rootParameters[0].Descriptor.RegisterSpace  = 0;
+    rootParameters[0].ShaderVisibility          = D3D12_SHADER_VISIBILITY_ALL;
+    // DrawParams (b1)
+    rootParameters[1].ParameterType            = D3D12_ROOT_PARAMETER_TYPE_32BIT_CONSTANTS;
+    rootParameters[1].Constants.Num32BitValues = 36;
+    rootParameters[1].Constants.ShaderRegister = 1;
+    rootParameters[1].Constants.RegisterSpace  = 0;
+    rootParameters[1].ShaderVisibility         = D3D12_SHADER_VISIBILITY_ALL;
+    // MaterialParams (t2)
+    rootParameters[2].ParameterType             = D3D12_ROOT_PARAMETER_TYPE_SRV;
+    rootParameters[2].Descriptor.ShaderRegister = 2;
+    rootParameters[2].Descriptor.RegisterSpace  = 0;
+    rootParameters[2].ShaderVisibility          = D3D12_SHADER_VISIBILITY_ALL;
+    // BRDFLUT (t10)
+    rootParameters[3].ParameterType                       = D3D12_ROOT_PARAMETER_TYPE_DESCRIPTOR_TABLE;
+    rootParameters[3].DescriptorTable.NumDescriptorRanges = 1;
+    rootParameters[3].DescriptorTable.pDescriptorRanges   = &range1;
+    rootParameters[3].ShaderVisibility                    = D3D12_SHADER_VISIBILITY_ALL;
+    // IrradianceMap (t16)
+    rootParameters[4].ParameterType                       = D3D12_ROOT_PARAMETER_TYPE_DESCRIPTOR_TABLE;
+    rootParameters[4].DescriptorTable.NumDescriptorRanges = 1;
+    rootParameters[4].DescriptorTable.pDescriptorRanges   = &range2;
+    rootParameters[4].ShaderVisibility                    = D3D12_SHADER_VISIBILITY_ALL;
+    // EnvironmentMap (t32)
+    rootParameters[5].ParameterType                       = D3D12_ROOT_PARAMETER_TYPE_DESCRIPTOR_TABLE;
+    rootParameters[5].DescriptorTable.NumDescriptorRanges = 1;
+    rootParameters[5].DescriptorTable.pDescriptorRanges   = &range3;
+    rootParameters[5].ShaderVisibility                    = D3D12_SHADER_VISIBILITY_ALL;
+
+    D3D12_STATIC_SAMPLER_DESC staticSamplers[2] = {};
+    // ClampedSampler (s4)
+    staticSamplers[0].Filter           = D3D12_FILTER_MIN_MAG_MIP_LINEAR;
+    staticSamplers[0].AddressU         = D3D12_TEXTURE_ADDRESS_MODE_CLAMP;
+    staticSamplers[0].AddressV         = D3D12_TEXTURE_ADDRESS_MODE_CLAMP;
+    staticSamplers[0].AddressW         = D3D12_TEXTURE_ADDRESS_MODE_CLAMP;
+    staticSamplers[0].MipLODBias       = D3D12_DEFAULT_MIP_LOD_BIAS;
+    staticSamplers[0].MaxAnisotropy    = 0;
+    staticSamplers[0].ComparisonFunc   = D3D12_COMPARISON_FUNC_LESS_EQUAL;
+    staticSamplers[0].MinLOD           = 0;
+    staticSamplers[0].MaxLOD           = 1;
+    staticSamplers[0].ShaderRegister   = 4;
+    staticSamplers[0].RegisterSpace    = 0;
+    staticSamplers[0].ShaderVisibility = D3D12_SHADER_VISIBILITY_PIXEL;
+    // UWrapSampler (s5)
+    staticSamplers[1].Filter           = D3D12_FILTER_MIN_MAG_MIP_LINEAR;
+    staticSamplers[1].AddressU         = D3D12_TEXTURE_ADDRESS_MODE_WRAP;
+    staticSamplers[1].AddressV         = D3D12_TEXTURE_ADDRESS_MODE_CLAMP;
+    staticSamplers[1].AddressW         = D3D12_TEXTURE_ADDRESS_MODE_CLAMP;
+    staticSamplers[1].MipLODBias       = D3D12_DEFAULT_MIP_LOD_BIAS;
+    staticSamplers[1].MaxAnisotropy    = 0;
+    staticSamplers[1].ComparisonFunc   = D3D12_COMPARISON_FUNC_LESS_EQUAL;
+    staticSamplers[1].MinLOD           = 0;
+    staticSamplers[1].MaxLOD           = D3D12_FLOAT32_MAX;
+    staticSamplers[1].ShaderRegister   = 5;
+    staticSamplers[1].RegisterSpace    = 0;
+    staticSamplers[1].ShaderVisibility = D3D12_SHADER_VISIBILITY_PIXEL;
+
+    D3D12_ROOT_SIGNATURE_DESC rootSigDesc = {};
+    rootSigDesc.NumParameters             = 6;
+    rootSigDesc.pParameters               = rootParameters;
+    rootSigDesc.NumStaticSamplers         = 2;
+    rootSigDesc.pStaticSamplers           = staticSamplers;
+    rootSigDesc.Flags                     = D3D12_ROOT_SIGNATURE_FLAG_ALLOW_INPUT_ASSEMBLER_INPUT_LAYOUT;
+
+    ComPtr<ID3DBlob> blob;
+    ComPtr<ID3DBlob> error;
+    CHECK_CALL(D3D12SerializeRootSignature(&rootSigDesc, D3D_ROOT_SIGNATURE_VERSION_1, &blob, &error));
+    CHECK_CALL(pRenderer->Device->CreateRootSignature(
+        0,                         // nodeMask
+        blob->GetBufferPointer(),  // pBloblWithRootSignature
+        blob->GetBufferSize(),     // blobLengthInBytes
+        IID_PPV_ARGS(ppRootSig))); // riid, ppvRootSignature
+}
+
+void CreateEnvironmentRootSig(DxRenderer* pRenderer, ID3D12RootSignature** ppRootSig)
+{
+    // Textures (t32)
+    D3D12_DESCRIPTOR_RANGE range            = {};
+    range.RangeType                         = D3D12_DESCRIPTOR_RANGE_TYPE_SRV;
+    range.NumDescriptors                    = gMaxIBLs;
+    range.BaseShaderRegister                = 32;
+    range.RegisterSpace                     = 0;
+    range.OffsetInDescriptorsFromTableStart = D3D12_DESCRIPTOR_RANGE_OFFSET_APPEND;
+
+    D3D12_ROOT_PARAMETER rootParameters[4] = {};
+    // SceneParams (b0)
+    rootParameters[0].ParameterType            = D3D12_ROOT_PARAMETER_TYPE_32BIT_CONSTANTS;
+    rootParameters[0].Constants.Num32BitValues = 17;
+    rootParameters[0].Constants.ShaderRegister = 0;
+    rootParameters[0].Constants.RegisterSpace  = 0;
+    rootParameters[0].ShaderVisibility         = D3D12_SHADER_VISIBILITY_ALL;
+    // Textures (t32)
+    rootParameters[1].ParameterType                       = D3D12_ROOT_PARAMETER_TYPE_DESCRIPTOR_TABLE;
+    rootParameters[1].DescriptorTable.NumDescriptorRanges = 1;
+    rootParameters[1].DescriptorTable.pDescriptorRanges   = &range;
+    rootParameters[1].ShaderVisibility                    = D3D12_SHADER_VISIBILITY_ALL;
+
+    D3D12_STATIC_SAMPLER_DESC staticSampler = {};
+    // Sampler0 (s1)
+    staticSampler.Filter           = D3D12_FILTER_MIN_MAG_MIP_LINEAR;
+    staticSampler.AddressU         = D3D12_TEXTURE_ADDRESS_MODE_CLAMP;
+    staticSampler.AddressV         = D3D12_TEXTURE_ADDRESS_MODE_CLAMP;
+    staticSampler.AddressW         = D3D12_TEXTURE_ADDRESS_MODE_CLAMP;
+    staticSampler.MipLODBias       = D3D12_DEFAULT_MIP_LOD_BIAS;
+    staticSampler.MaxAnisotropy    = 0;
+    staticSampler.ComparisonFunc   = D3D12_COMPARISON_FUNC_LESS_EQUAL;
+    staticSampler.MinLOD           = 0;
+    staticSampler.MaxLOD           = 1;
+    staticSampler.ShaderRegister   = 1;
+    staticSampler.RegisterSpace    = 0;
+    staticSampler.ShaderVisibility = D3D12_SHADER_VISIBILITY_PIXEL;
+
+    D3D12_ROOT_SIGNATURE_DESC rootSigDesc = {};
+    rootSigDesc.NumParameters             = 2;
+    rootSigDesc.pParameters               = rootParameters;
+    rootSigDesc.NumStaticSamplers         = 1;
+    rootSigDesc.pStaticSamplers           = &staticSampler;
+    rootSigDesc.Flags                     = D3D12_ROOT_SIGNATURE_FLAG_ALLOW_INPUT_ASSEMBLER_INPUT_LAYOUT;
+
+    ComPtr<ID3DBlob> blob;
+    ComPtr<ID3DBlob> error;
+    CHECK_CALL(D3D12SerializeRootSignature(&rootSigDesc, D3D_ROOT_SIGNATURE_VERSION_1, &blob, &error));
+    CHECK_CALL(pRenderer->Device->CreateRootSignature(
+        0,                         // nodeMask
+        blob->GetBufferPointer(),  // pBloblWithRootSignature
+        blob->GetBufferSize(),     // blobLengthInBytes
+        IID_PPV_ARGS(ppRootSig))); // riid, ppvRootSignature
+}
+
+void CreateDescriptorHeap(
+    DxRenderer*            pRenderer,
+    ID3D12DescriptorHeap** ppHeap)
+{
+    D3D12_DESCRIPTOR_HEAP_DESC desc = {};
+    desc.Type                       = D3D12_DESCRIPTOR_HEAP_TYPE_CBV_SRV_UAV;
+    desc.NumDescriptors             = 256;
+    desc.Flags                      = D3D12_DESCRIPTOR_HEAP_FLAG_SHADER_VISIBLE;
+
+    CHECK_CALL(pRenderer->Device->CreateDescriptorHeap(
+        &desc,
+        IID_PPV_ARGS(ppHeap)));
+}
+*/
