+++ projects/pbr/252_pbr_material_properties_d3d12/CMakeLists.txt
@@ -1,48 +1,40 @@
 cmake_minimum_required(VERSION 3.0)
 
-project(203_pbr_camera_d3d12)
+project(252_pbr_material_properties_d3d12)
 
 add_executable(
-    203_pbr_camera_d3d12
-    203_pbr_camera_d3d12.cpp
+    252_pbr_material_properties_d3d12
+    252_pbr_material_properties_d3d12.cpp
     ${GREX_PROJECTS_COMMON_DIR}/config.h
     ${GREX_PROJECTS_COMMON_DIR}/dx_renderer.h
     ${GREX_PROJECTS_COMMON_DIR}/dx_renderer.cpp
     ${GREX_PROJECTS_COMMON_DIR}/window.h
     ${GREX_PROJECTS_COMMON_DIR}/window.cpp
-    ${GREX_PROJECTS_COMMON_DIR}/tri_mesh.h
-    ${GREX_PROJECTS_COMMON_DIR}/tri_mesh.cpp
     ${GREX_PROJECTS_COMMON_DIR}/bitmap.h
     ${GREX_PROJECTS_COMMON_DIR}/bitmap.cpp
-    ${GREX_THIRD_PARTY_DIR}/MikkTSpace/mikktspace.h
-    ${GREX_THIRD_PARTY_DIR}/MikkTSpace/mikktspace.c
+    ${GREX_PROJECTS_COMMON_DIR}/tri_mesh.h
+    ${GREX_PROJECTS_COMMON_DIR}/tri_mesh.cpp
     ${IMGUI_D3D12_FILES}
 )
 
-set_target_properties(203_pbr_camera_d3d12 PROPERTIES FOLDER "pbr")
+set_target_properties(252_pbr_material_properties_d3d12 PROPERTIES FOLDER "pbr")
 
 target_compile_definitions(
-    203_pbr_camera_d3d12
-    PUBLIC TRIMESH_USE_MIKKTSPACE
-)
-
-target_compile_definitions(
-    203_pbr_camera_d3d12
+    252_pbr_material_properties_d3d12
     PUBLIC ENABLE_IMGUI_D3D12
 )
 
 target_include_directories(
-    203_pbr_camera_d3d12
+    252_pbr_material_properties_d3d12
     PUBLIC ${GREX_PROJECTS_COMMON_DIR}
            ${GREX_THIRD_PARTY_DIR}/glm
-           ${GREX_THIRD_PARTY_DIR}/tinyobjloader
            ${GREX_THIRD_PARTY_DIR}/stb
-           ${GREX_THIRD_PARTY_DIR}/MikkTSpace
+           ${GREX_THIRD_PARTY_DIR}/tinyobjloader
            ${IMGUI_INC_DIR}
 )
 
 target_link_libraries(
-    203_pbr_camera_d3d12
+    252_pbr_material_properties_d3d12
     PUBLIC glfw
            d3d12
            dxgi
