+++ projects/pbr/252_pbr_material_properties_vulkan/252_pbr_material_properties_vulkan.cpp
@@ -170,7 +170,7 @@ int main(int argc, char** argv)
     std::vector<uint32_t> spirvVS;
     std::vector<uint32_t> spirvFS;
     {
-        std::string shaderSource = LoadString("projects/253_254_pbr_material_properties/shaders.hlsl");
+        std::string shaderSource = LoadString("projects/252_pbr_material_properties/shaders.hlsl");
 
         std::string errorMsg;
         HRESULT     hr = CompileHLSL(shaderSource, "vsmain", "vs_6_0", &spirvVS, &errorMsg);
