+++ projects/pbr/252_pbr_material_properties_vulkan/CMakeLists.txt
@@ -1,66 +1,59 @@
 cmake_minimum_required(VERSION 3.0)
 
-project(204_pbr_camera_vulkan)
+project(252_pbr_material_properties_vulkan)
 
 add_executable(
-    204_pbr_camera_vulkan
-    204_pbr_camera_vulkan.cpp
+    252_pbr_material_properties_vulkan
+    252_pbr_material_properties_vulkan.cpp
     ${GREX_PROJECTS_COMMON_DIR}/config.h
     ${GREX_PROJECTS_COMMON_DIR}/vk_renderer.h
     ${GREX_PROJECTS_COMMON_DIR}/vk_renderer.cpp
     ${GREX_PROJECTS_COMMON_DIR}/window.h
     ${GREX_PROJECTS_COMMON_DIR}/window.cpp
-    ${GREX_PROJECTS_COMMON_DIR}/tri_mesh.h
-    ${GREX_PROJECTS_COMMON_DIR}/tri_mesh.cpp
     ${GREX_PROJECTS_COMMON_DIR}/bitmap.h
     ${GREX_PROJECTS_COMMON_DIR}/bitmap.cpp
-    ${GREX_THIRD_PARTY_DIR}/MikkTSpace/mikktspace.h
-    ${GREX_THIRD_PARTY_DIR}/MikkTSpace/mikktspace.c
+    ${GREX_PROJECTS_COMMON_DIR}/tri_mesh.h
+    ${GREX_PROJECTS_COMMON_DIR}/tri_mesh.cpp
     ${GREX_THIRD_PARTY_DIR}/glslang/StandAlone/resource_limits_c.cpp
     ${GREX_THIRD_PARTY_DIR}/glslang/StandAlone/ResourceLimits.cpp
     ${IMGUI_VULKAN_FILES}
 )
 
-set_target_properties(204_pbr_camera_vulkan PROPERTIES FOLDER "pbr")
+set_target_properties(252_pbr_material_properties_vulkan PROPERTIES FOLDER "pbr")
 
 target_compile_definitions(
-    204_pbr_camera_vulkan
-    PUBLIC TRIMESH_USE_MIKKTSPACE
-)
-
-target_compile_definitions(
-    204_pbr_camera_vulkan
+    252_pbr_material_properties_vulkan
     PUBLIC ENABLE_IMGUI_VULKAN
 )
 
 target_include_directories(
-    204_pbr_camera_vulkan
+    252_pbr_material_properties_vulkan
     PUBLIC ${GREX_PROJECTS_COMMON_DIR}
            ${GREX_THIRD_PARTY_DIR}/glslang # This needs to come before ${VULKAN_INCLUDE_DIR}
 		   ${VULKAN_INCLUDE_DIR}
            ${GREX_THIRD_PARTY_DIR}/VulkanMemoryAllocator/include
            ${GREX_THIRD_PARTY_DIR}/glm
-           ${GREX_THIRD_PARTY_DIR}/tinyobjloader
            ${GREX_THIRD_PARTY_DIR}/stb
-           ${GREX_THIRD_PARTY_DIR}/MikkTSpace
+           ${GREX_THIRD_PARTY_DIR}/tinyobjloader
            ${IMGUI_INC_DIR}
 )
 
 target_link_libraries(
-    204_pbr_camera_vulkan
+    252_pbr_material_properties_vulkan
     PUBLIC glfw
            glslang
            SPIRV
            dxcompiler
 )
+
 if(WIN32)
     target_compile_definitions(
-        204_pbr_camera_vulkan
+        252_pbr_material_properties_vulkan
         PUBLIC VK_USE_PLATFORM_WIN32_KHR
     )
 
     target_link_libraries(
-        204_pbr_camera_vulkan
+        252_pbr_material_properties_vulkan
         PUBLIC "${VULKAN_LIBRARY_DIR}/vulkan-1.lib"
     )
 endif()
